### 2022-05-26 #4429 @arfon
@adam2392 – JOSS can't compile a paper that's on a remote branch (`adam2392:joss`) sorry. Can you merge it please?
