### 2022-05-26 7202a3d
[PAPER] JOSS for v0.1/0.2 ICLabel port (#61)
* INitial commit for joss paper

* iNitial commit

* Add paper

* Apply suggestions from code review

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Add paper

* Apply suggestions from code review

* Adding to manifest and updated format

* Mod sentence

* Update paper/paper.md

* Adding coauthor

* Update paper/paper.md

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Add JOSS gh action

* Add joss

* Format yml

* Try separate yaml?

* Add multiple affiliation

* Add multiple affiliation

* fix syntax
[…]
