### 2022-05-30 #4429 @Kevin-Mattheus-Moerman
@adam2392 you call `@editorialbot generate pdf`
