### 2022-06-04 b468fb1
remove old sanity-check
