+++ mne_icalabel/iclabel/features.py
@@ -314,8 +314,6 @@ def _eeg_rpsd_compute_psdmed(
         elif isinstance(inst, BaseEpochs):
             temp = np.hstack([icaact[it, index[:, k], :] for k in range(index.shape[-1])])
             temp = temp.reshape(index.shape[0], len(inst), order="F")
-        else:
-            raise RuntimeError  # should never happen
         temp = (temp[:, subset].T * window).T
         temp = np.fft.fft(temp, n_points, axis=0)
         temp = temp * np.conjugate(temp)
