### 2022-06-04 0c66170
add test for constants on long epo
