+++ mne_icalabel/iclabel/features.py
@@ -312,8 +312,8 @@ def _eeg_rpsd_compute_psdmed(
             temp = np.hstack([icaact[it, index[:, k]] for k in range(index.shape[-1])])
             temp = temp.reshape(*index.shape, order="F")
         elif isinstance(inst, BaseEpochs):
-            temp = np.hstack([icaact[it, index[:, k], :] for k in range(index.shape[-1])])
-            temp = temp.reshape(index.shape[0], len(inst), order="F")
+            temp = np.vstack([icaact[it, index[:, k], :] for k in range(index.shape[-1])])
+            temp = temp.reshape(index.shape[0], index.shape[1] * len(inst), order="F")
         temp = (temp[:, subset].T * window).T
         temp = np.fft.fft(temp, n_points, axis=0)
         temp = temp * np.conjugate(temp)
