### 2022-06-05 c19a567
add run test on different data shapes
