### 2022-06-06 bc153f1
[BUG] Fix the PSD features of ICLabel with epochs of more than 1 second (#64)
* remove old sanity-check

* add test for constants on long epo

* fix shape issue

* add test

* fix typo

* don't redefine reader [ci skip]

* double check vstack vs hstack and add comment

* add run test on different data shapes

* add entry to changelog

* update hash and simplify test
