+++ mne_icalabel/iclabel/tests/test_features.py
@@ -32,6 +32,7 @@ raw_eeglab_path = dataset_path / "datasets/sample-raw.set"
 raw_short_eeglab_path = dataset_path / "datasets/sample-short-raw.set"
 raw_very_short_eeglab_path = dataset_path / "datasets/sample-very-short-raw.set"
 epo_eeglab_path = dataset_path / "datasets/sample-epo.set"
+epo_long_eeglab_path = dataset_path / "datasets/sample-long-epo.set"
 
 # ICA activation matrix for raw/epochs
 raw_icaact_eeglab_path = dataset_path / "icaact/icaact-raw.mat"
@@ -50,6 +51,9 @@ psd_raw_path = dataset_path / "psd/psd-raw.mat"
 psd_constants_epo_path = dataset_path / "psd/constants-epo.mat"
 psd_steps_epo_path = dataset_path / "psd/psd-step-by-step-epo.mat"
 psd_epo_path = dataset_path / "psd/psd-epo.mat"
+psd_constants_long_epo_path = dataset_path / "psd/constants-long-epo.mat"
+psd_steps_long_epo_path = dataset_path / "psd/psd-step-by-step-long-epo.mat"
+psd_long_epo_path = dataset_path / "psd/psd-long-epo.mat"
 
 # Autocorrelations
 autocorr_raw_path = dataset_path / "autocorr/autocorr-raw.mat"
@@ -194,43 +198,21 @@ def test_eeg_topoplot(file, eeglab_result_file):
 
 
 # ----------------------------------------------------------------------------
-def test_eeg_rpsd_constants():
+@pytest.mark.parametrize(
+    "fname, constants_fname, type_",
+    (
+        (raw_eeglab_path, psd_constants_raw_path, "raw"),
+        (epo_eeglab_path, psd_constants_epo_path, "epo"),
+        (epo_long_eeglab_path, psd_constants_long_epo_path, "epo"),
+    ),
+)
+def test_eeg_rpsd_constants(fname, constants_fname, type_):
     """Test _eeg_rpsd_constants function."""
-    # Raw --------------------------------------------------------------------
-    raw = read_raw(raw_eeglab_path, preload=True)
-    ica = read_ica_eeglab(raw_eeglab_path)
-    ncomp, nfreqs, n_points, nyquist, index, window, subset = _eeg_rpsd_constants(raw, ica)
+    inst = reader[type_](fname, **kwargs[type_])
+    ica = read_ica_eeglab(fname)
+    ncomp, nfreqs, n_points, nyquist, index, window, subset = _eeg_rpsd_constants(inst, ica)
 
-    constants_eeglab = loadmat(psd_constants_raw_path)["constants"][0, 0]
-    ncomp_eeglab = constants_eeglab["ncomp"][0, 0]
-    nfreqs_eeglab = constants_eeglab["nfreqs"][0, 0]
-    n_points_eeglab = constants_eeglab["n_points"][0, 0]
-    nyquist_eeglab = constants_eeglab["nyquist"][0, 0]
-    index_eeglab = constants_eeglab["index"]
-    window_eeglab = constants_eeglab["window"]
-    subset_eeglab = constants_eeglab["subset"]
-
-    assert ncomp == ncomp_eeglab
-    assert nfreqs == nfreqs_eeglab
-    assert n_points == n_points_eeglab
-    assert nyquist == nyquist_eeglab
-    assert np.allclose(index, index_eeglab - 1)
-
-    # window and subset are not squeezed in matlab
-    assert window_eeglab.shape[0] == 1
-    assert np.allclose(window, window_eeglab[0, :])
-
-    # for subsets, compare if the same elements are in both
-    assert subset_eeglab.shape[0] == 1
[…]
