+++ mne_icalabel/datasets/icalabel/tests/test_dataset_icalabel.py
@@ -1,14 +1,12 @@
-import os.path as op
-
 import mne_icalabel
 
 
 def test_dataset_tapping_group():
     datapath = mne_icalabel.datasets.icalabel.data_path()
-    assert op.isdir(datapath)
-    assert op.isdir(op.join(datapath, "iclabel"))
+    assert datapath.is_dir()
+    assert (datapath / "iclabel").is_dir()
 
     # First pass downloaded, check that second pass of access works
     datapath = mne_icalabel.datasets.icalabel.data_path()
-    assert op.isdir(datapath)
-    assert op.isdir(op.join(datapath, "iclabel"))
+    assert datapath.is_dir()
+    assert (datapath / "iclabel").is_dir()
