### 2022-06-10 #4429 @jasmainak
Hi @emdupre, sorry for the late response as I’m out of town currently. I’d
be happy to review!

I do have a COI with Adam Li on the mne-bids paper which was published in
JOSS 3 years ago. Does that make me ineligible to review?

On Thu 9 Jun 2022 at 20:08, Martina G. Vilas ***@***.***>
wrote:

> Hi @emdupre <https://github.com/emdupre> 👋 Thank you for the invitation.
> Unfortunately, at the moment I'm lacking the time to review due to other
> commitments. I hope to be available for other reviews in the near future.
>
> The article looks very interesting, all the best to the authors!
>
