### 2022-06-13 #4429 @adam2392
> Hi @emdupre, sorry for the late response as I’m out of town currently. I’d be happy to review! I do have a COI with Adam Li on the mne-bids paper which was published in JOSS 3 years ago. Does that make me ineligible to review?

I actually wasn't a co-author on that publication :p

However, I did work with @jasmainak on the MNE-BIDs package a lot post-JOSS publication.
