### 2022-06-15 1a6df47
Adding the GUI with basic Python API
