+++ mne_icalabel/gui/_label_components_2.py
@@ -1,0 +1,186 @@
+import platform
+
+from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
+from matplotlib.figure import Figure
+from qtpy.QtCore import Qt, Slot
+from qtpy.QtWidgets import (
+    QAbstractItemView,
+    QButtonGroup,
+    QGridLayout,
+    QListWidget,
+    QMainWindow,
+    QPushButton,
+    QVBoxLayout,
+    QWidget,
+)
+
+_CH_MENU_WIDTH = 30 if platform.system() == "Windows" else 10
+
+
+class TopomapFig(FigureCanvasQTAgg):
+    """Topographic map figure widget."""
+
+    def __init__(self, width=4, height=4, dpi=300):
+        fig = Figure(figsize=(width, height), dpi=dpi)
+        ax = fig.subplots()
+        fig.subplots_adjust(bottom=0, left=0, right=1, top=1, wspace=0, hspace=0)
+        # clean up excess plot text, invert
+        ax.invert_yaxis()
+        ax.set_xticks([])
+        ax.set_yticks([])
+        super().__init__(fig)
+
+    def change_ic(self, ica, idx):
+        pass
+
+
+class PowerSpectralDensityFig(FigureCanvasQTAgg):
+    """PSD figure widget."""
+
+    def __init__(self, width=4, height=4, dpi=300):
+        fig = Figure(figsize=(width, height), dpi=dpi)
+        ax = fig.subplots()
+        fig.subplots_adjust(bottom=0, left=0, right=1, top=1, wspace=0, hspace=0)
+        # clean up excess plot text, invert
+        ax.set_xticks([])
+        ax.set_yticks([])
+        super().__init__(fig)
+
+    def change_ic(self, ica, idx):
+        pass
+
+
+class TimeSeriesFig(FigureCanvasQTAgg):
+    """Time-series figure widget."""
+
+    def __init__(self, width=4, height=4, dpi=300):
+        fig = Figure(figsize=(width, height), dpi=dpi)
+        ax = fig.subplots()
+        fig.subplots_adjust(bottom=0, left=0, right=1, top=1, wspace=0, hspace=0)
+        # clean up excess plot text, invert
+        ax.set_xticks([])
+        ax.set_yticks([])
+        super().__init__(fig)
+
+    def change_ic(self, ica, inst, idx):
+        pass
+
+
+# TODO: Maybe that should inherit from a QGroupBox?
+class Labels(QWidget):
+    """Widget with labels as push buttons.
+
+    Only one of the labels can be selected at once.
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        labels = [
+            "Brain",
+            "Eye",
+            "Heart",
+            "Muscle",
+            "Channel Noise",
+            "Line Noise",
+            "Other",
+        ]
+
+        self.buttonGroup = QButtonGroup()
+        self.buttonGroup.setExclusive(True)
+        layout = QVBoxLayout()
+        for k, label in enumerate(labels):
+            pushButton = Labels.create_pushButton(label)
+            self.buttonGroup.addButton(pushButton, k)
+            layout.addWidget(pushButton)
+        self.setLayout(layout)
+        # connect signal to slots
+        self.buttonGroup.buttonClicked.connect(self.pushButton_clicked)
+
+    @staticmethod
+    def create_pushButton(label):
+        pushButton = QPushButton()
+        pushButton.setObjectName(f"pushButton_{label.lower().replace(' ', '_')}")
+        pushButton.setText(label)
+        pushButton.setCheckable(True)
+        pushButton.setChecked(False)
+        pushButton.setEnabled(False)
[…]
