### 2022-06-16 d0b1c0a
fix channel selection for topoplot
