+++ mne_icalabel/iclabel/tests/test_utils.py
@@ -40,7 +40,7 @@ def test_loc(file, eeglab_result_file):
     when loading the datasets."""
     type_ = str(Path(file).stem)[-3:]
     inst = reader[type_](file, **kwargs[type_])
-    rd, th = _mne_to_eeglab_locs(inst)
+    rd, th = _mne_to_eeglab_locs(inst, picks=inst.ch_names)
     eeglab_loc = loadmat(eeglab_result_file)["loc"][0, 0]
     eeglab_rd = eeglab_loc["rd"]
     eeglab_th = eeglab_loc["th"]
