+++ mne_icalabel/iclabel/tests/test_label_components.py
@@ -1,15 +1,15 @@
 import numpy as np
 import pytest
 from mne import create_info, pick_types, make_fixed_length_epochs
-from mne.datasets import sample
+from mne.datasets import testing
 from mne.io import RawArray, read_raw
 from mne.preprocessing import ICA
 
 from mne_icalabel.iclabel import iclabel_label_components
 
-directory = sample.data_path() / "MEG" / "sample"
-raw = read_raw(directory / "sample_audvis_raw.fif", preload=False)
-raw.crop(0, 100).pick_types(eeg=True, exclude="bads")
+directory = testing.data_path() / "MEG" / "sample"
+raw = read_raw(directory / "sample_audvis_trunc_raw.fif", preload=False)
+raw.pick_types(eeg=True, exclude=[])
 raw.load_data()
 # preprocess
 raw.filter(l_freq=1.0, h_freq=100.0)
@@ -21,19 +21,17 @@ raw.set_eeg_reference("average")
     "inst, exclude",
     (
         (raw, "bads"),
-        (raw.copy().crop(0, 10), "bads"),
+        (raw.copy().crop(0, 8), "bads"),
         (raw.copy().crop(0, 1), "bads"),
         (make_fixed_length_epochs(raw, duration=0.5, preload=True), "bads"),
         (make_fixed_length_epochs(raw, duration=1, preload=True), "bads"),
         (make_fixed_length_epochs(raw, duration=5, preload=True), "bads"),
-        (make_fixed_length_epochs(raw, duration=10, preload=True), "bads"),
         (raw, []),
-        (raw.copy().crop(0, 10), []),
+        (raw.copy().crop(0, 8), []),
         (raw.copy().crop(0, 1), []),
         (make_fixed_length_epochs(raw, duration=0.5, preload=True), []),
         (make_fixed_length_epochs(raw, duration=1, preload=True), []),
         (make_fixed_length_epochs(raw, duration=5, preload=True), []),
-        (make_fixed_length_epochs(raw, duration=10, preload=True), []),
     ),
 )
 def test_label_components(inst, exclude):
