### 2022-06-16 ab090dd
Fix ICLabel topographic features on ICA fitted with channel selection (#68)
* fix channel selection for topoplot

* fix style

* fix tests

* add test

* use testing instead of sample dataset

* fix style

* add entry to changelog

* add gh id
