+++ mne_icalabel/iclabel/utils.py
@@ -5,7 +5,7 @@ from mne.io import BaseRaw
 from numpy.typing import ArrayLike, NDArray
 
 
-def _mne_to_eeglab_locs(raw: BaseRaw) -> Tuple[NDArray[float], NDArray[float]]:
+def _mne_to_eeglab_locs(raw: BaseRaw, picks: List[str]) -> Tuple[NDArray[float], NDArray[float]]:
     """Obtain EEGLab-like spherical coordinate from EEG channel positions.
 
     TODO: @JACOB:
@@ -18,6 +18,8 @@ def _mne_to_eeglab_locs(raw: BaseRaw) -> Tuple[NDArray[float], NDArray[float]]:
     raw : mne.io.BaseRaw
         Instance of raw object with a `mne.montage.DigMontage` set with
         ``n_channels`` channel positions.
+    picks : list of str
+        List of channel names to include.
 
     Returns
     -------
@@ -44,7 +46,7 @@ def _mne_to_eeglab_locs(raw: BaseRaw) -> Tuple[NDArray[float], NDArray[float]]:
         return azimuth, elevation, r
 
     # get the channel position dictionary
-    montage = raw.get_montage()
+    montage = raw.copy().pick_channels(picks, ordered=True).get_montage()
     positions = montage.get_positions()
     ch_pos = positions["ch_pos"]
 
