+++ mne_icalabel/iclabel/tests/test_label_components.py
@@ -1,15 +1,15 @@
 import numpy as np
 import pytest
-from mne import create_info, make_fixed_length_epochs
-from mne.datasets import sample
+from mne import create_info, make_fixed_length_epochs, pick_types
+from mne.datasets import testing
 from mne.io import RawArray, read_raw
 from mne.preprocessing import ICA
 
 from mne_icalabel.iclabel import iclabel_label_components
 
-directory = sample.data_path() / "MEG" / "sample"
-raw = read_raw(directory / "sample_audvis_raw.fif", preload=False)
-raw.crop(0, 100).pick_types(eeg=True, exclude="bads")
+directory = testing.data_path() / "MEG" / "sample"
+raw = read_raw(directory / "sample_audvis_trunc_raw.fif", preload=False)
+raw.pick_types(eeg=True, exclude=[])
 raw.load_data()
 # preprocess
 raw.filter(l_freq=1.0, h_freq=100.0)
@@ -18,21 +18,27 @@ raw.set_eeg_reference("average")
 
 @pytest.mark.filterwarnings("ignore::RuntimeWarning")
 @pytest.mark.parametrize(
-    "inst",
+    "inst, exclude",
     (
-        raw,
-        raw.copy().crop(0, 10),
-        raw.copy().crop(0, 1),
-        make_fixed_length_epochs(raw, duration=0.5),
-        make_fixed_length_epochs(raw, duration=1),
-        make_fixed_length_epochs(raw, duration=5),
-        make_fixed_length_epochs(raw, duration=10),
+        (raw, "bads"),
+        (raw.copy().crop(0, 8), "bads"),
+        (raw.copy().crop(0, 1), "bads"),
+        (make_fixed_length_epochs(raw, duration=0.5, preload=True), "bads"),
+        (make_fixed_length_epochs(raw, duration=1, preload=True), "bads"),
+        (make_fixed_length_epochs(raw, duration=5, preload=True), "bads"),
+        (raw, []),
+        (raw.copy().crop(0, 8), []),
+        (raw.copy().crop(0, 1), []),
+        (make_fixed_length_epochs(raw, duration=0.5, preload=True), []),
+        (make_fixed_length_epochs(raw, duration=1, preload=True), []),
+        (make_fixed_length_epochs(raw, duration=5, preload=True), []),
     ),
 )
-def test_label_components(inst):
+def test_label_components(inst, exclude):
     """Check that label_components does not raise on various data shapes."""
+    picks = pick_types(raw.info, eeg=True, exclude=exclude)
     ica = ICA(n_components=5, method="picard", fit_params=dict(ortho=False, extended=True))
-    ica.fit(inst)
+    ica.fit(inst, picks=picks)
     labels = iclabel_label_components(inst, ica)
     assert labels.shape == (ica.n_components_, 7)
 
