### 2022-06-16 #4429 @adam2392
Hi @emdupre are there any updates here, or action items needed from us?

Also just wanted to see if there was any insight into what we should expect and when for the JOSS process? Is there a summary you can point me to perhaps? Thanks!
