### 2022-06-16 #4429 @emdupre
👋 Hi @TomDonoghue, @rdgao, @ixjlyons  !

Would you be willing to review _MNE-ICALabel: Automatically annotating ICA components with ICLabel in Python_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/mne-tools/mne-icalabel) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.04429/joss.04429/10.21105.joss.04429.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you !
