### 2022-06-18 #4484 @editorialbot
[REVIEW]: MNE-ICALabel: Automatically annotating ICA components with ICLabel in Python
**Submitting author:** @adam2392 (<a href="http://orcid.org/0000-0001-8421-365X">Adam Li</a>)
**Repository:** https://github.com/mne-tools/mne-icalabel
**Branch with paper.md** (empty if default branch): main
**Version:** v0.4
