### 2022-06-18 #4484 @emdupre
Hi  @hvgazula and @TomDonoghue 👋 Thanks again for agreeing to review this submission ! The review will take place in this issue, and you can generate your individual reviewer checklists by asking `editorialbot` directly with `@\editorialbot generate my checklist`.

In working through the checklist, you're likely to have specific feedback on _MNE-ICALabel_. Whenever possible, please open relevant issues on the [linked software repository](https://github.com/mne-tools/mne-icalabel) (and cross-link them with this issue) rather than discussing them here. This helps to make sure that feedback is translated into actionable items to improve the software !

If you aren't sure how to get started, please see the [Reviewing for JOSS guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) -- and, of course, feel free to ping me with any questions !
