### 2022-06-27 cb05faa
Adding testing script
