+++ mne_icalabel/gui/_label_components.py
@@ -11,21 +11,16 @@ import platform
 from matplotlib.backends.backend_qt5agg import FigureCanvas
 from matplotlib.figure import Figure
 from mne.preprocessing import ICA
-from mne.viz.backends.renderer import _get_renderer
+from mne.viz.utils import safe_event
 from qtpy import QtGui
-from qtpy.QtCore import Signal, Slot
+from qtpy.QtCore import Slot
 from qtpy.QtWidgets import (
     QAbstractItemView,
-    QComboBox,
     QGridLayout,
     QHBoxLayout,
-    QLabel,
     QListView,
     QMainWindow,
     QMessageBox,
-    QPlainTextEdit,
-    QPushButton,
-    QSlider,
     QVBoxLayout,
     QWidget,
 )
@@ -39,6 +34,7 @@ from qtpy.QtWidgets import (
 _CH_MENU_WIDTH = 30 if platform.system() == "Windows" else 10
 
 
+# TODO: remove
 def _make_topo_plot(width=4, height=4, dpi=300):
     """Make subplot for the topomap."""
     fig = Figure(figsize=(width, height), dpi=dpi)
@@ -53,6 +49,7 @@ def _make_topo_plot(width=4, height=4, dpi=300):
     return canvas, fig
 
 
+# TODO: remove
 def _make_ts_plot(width=4, height=4, dpi=300):
     """Make subplot for the component time-series."""
     fig = Figure(figsize=(width, height), dpi=dpi)
@@ -60,13 +57,13 @@ def _make_ts_plot(width=4, height=4, dpi=300):
     ax = fig.subplots()
     fig.subplots_adjust(bottom=0, left=0, right=1, top=1, wspace=0, hspace=0)
     ax.set_facecolor("k")
-    # clean up excess plot text, invert
-    ax.invert_yaxis()
+    # clean up excess plot text
     ax.set_xticks([])
     ax.set_yticks([])
     return canvas, fig
 
 
+# TODO: remove
 def _make_spectrum_plot(width=4, height=4, dpi=300):
     """Make subplot for the spectrum."""
     fig = Figure(figsize=(width, height), dpi=dpi)
@@ -74,11 +71,56 @@ def _make_spectrum_plot(width=4, height=4, dpi=300):
     ax = fig.subplots()
     fig.subplots_adjust(bottom=0, left=0, right=1, top=1, wspace=0, hspace=0)
     ax.set_facecolor("k")
-    # clean up excess plot text, invert
-    ax.invert_yaxis()
+    # clean up excess plot text
     ax.set_xticks([])
     ax.set_yticks([])
     return canvas, fig
+
+
+class TimeSeriesFig(FigureCanvas):
+    """Spectrum map widget."""
+
+    def __init__(self, width=4, height=4, dpi=300):
+        """Make subplot for the spectrum."""
+        fig = Figure(figsize=(width, height), dpi=dpi)
+        canvas = FigureCanvas(fig)
+        ax = fig.subplots()
+        fig.subplots_adjust(bottom=0, left=0, right=1, top=1, wspace=0, hspace=0)
+        ax.set_facecolor("k")
+        # clean up excess plot text
+        ax.set_xticks([])
+        ax.set_yticks([])
+        super().__init__(fig)
+
+
+class SpectrumFig(FigureCanvas):
+    """Spectrum map widget."""
+
+    def __init__(self, width=4, height=4, dpi=300):
+        """Make subplot for the spectrum."""
+        fig = Figure(figsize=(width, height), dpi=dpi)
+        canvas = FigureCanvas(fig)
+        ax = fig.subplots()
+        fig.subplots_adjust(bottom=0, left=0, right=1, top=1, wspace=0, hspace=0)
[…]
