+++ mne_icalabel/__init__.py
@@ -5,6 +5,6 @@
 #
 # License: BSD (3-clause)
 
-__version__ = "0.2dev0"
+__version__ = "0.2"
 
 from .label_components import label_components  # noqa: F401
