### 2022-06-30 058824d
Adding whatsnew for maint/v0.2
