### 2022-07-06 8dfdb8e
add entry-point to run the GUI and app.exec() to start the Qt event loop
