+++ setup.cfg
@@ -44,6 +44,10 @@ install_requires =
 packages = find:
 include_package_data = True
 
+[options.entry_points]
+console_scripts =
+    mne_gui_ic_annotation = mne_icalabel.commands.mne_gui_ic_annotation:main
+
 # Building package
 [bdist_wheel]
 universal = true
