### 2022-07-10 59d23a5
add saving of the selected labels to self.saved_labels and add a reset button
