### 2022-07-11 b68d584
remove ignore warnings from tests
