+++ mne_icalabel/tests/test_label_components.py
@@ -14,7 +14,6 @@ raw.filter(l_freq=1.0, h_freq=100.0)
 raw.set_eeg_reference("average")
 
 
-@pytest.mark.filterwarnings("ignore::RuntimeWarning")
 @pytest.mark.parametrize("n_components", (5, 15))
 def test_label_components(n_components):
     """Simple test to check that label_components runs without raising."""
