+++ mne_icalabel/iclabel/tests/test_features.py
@@ -71,7 +71,6 @@ kwargs = {"raw": dict(preload=True), "epo": dict()}
 
 
 # ----------------------------------------------------------------------------
-@pytest.mark.filterwarnings("ignore::RuntimeWarning")
 @pytest.mark.parametrize(
     "file, psd_constant_file, eeglab_feature_file",
     [
@@ -147,7 +146,6 @@ def test_compute_ica_activations(file, eeglab_result_file):
 
 
 # ----------------------------------------------------------------------------
-@pytest.mark.filterwarnings("ignore::RuntimeWarning")
 @pytest.mark.parametrize(
     "file, eeglab_result_file",
     [(raw_eeglab_path, raw_topo1_path), (epo_eeglab_path, epo_topo1_path)],
@@ -172,7 +170,6 @@ def test_topoplotFast(file, eeglab_result_file):
     assert np.allclose(topo1, topo1_eeglab, equal_nan=True)
 
 
-@pytest.mark.filterwarnings("ignore::RuntimeWarning")
 @pytest.mark.parametrize(
     "file, eeglab_result_file",
     [
