### 2022-07-11 5aac696
fix tests for RuntimeWarning because of ICA method and instance BP filter
