+++ mne_icalabel/iclabel/tests/test_label_components.py
@@ -12,7 +12,9 @@ raw = read_raw(directory / "sample_audvis_trunc_raw.fif", preload=False)
 raw.pick_types(eeg=True, exclude=[])
 raw.load_data()
 # preprocess
-raw.filter(l_freq=1.0, h_freq=100.0)
+with raw.info._unlock():  # fake filtering, testing dataset is filtered between [0.1, 80] Hz
+    raw.info["highpass"] = 1.
+    raw.info["lowpass"] = 100.
 raw.set_eeg_reference("average")
 
 
