+++ mne_icalabel/tests/test_label_components.py
@@ -10,14 +10,16 @@ raw = read_raw(directory / "sample_audvis_trunc_raw.fif", preload=False)
 raw.pick_types(eeg=True, exclude="bads")
 raw.load_data()
 # preprocess
-raw.filter(l_freq=1.0, h_freq=100.0)
+with raw.info._unlock():  # fake filtering, testing dataset is filtered between [0.1, 80] Hz
+    raw.info["highpass"] = 1.
+    raw.info["lowpass"] = 100.
 raw.set_eeg_reference("average")
 
 
 @pytest.mark.parametrize("n_components", (5, 15))
 def test_label_components(n_components):
     """Simple test to check that label_components runs without raising."""
-    ica = ICA(n_components=n_components, method="picard")
+    ica = ICA(n_components=n_components, method="infomax", fit_params=dict(extended=True))
     ica.fit(raw)
     labels = label_components(raw, ica, method="iclabel")
     assert isinstance(labels, dict)
