### 2022-07-11 29e8185
ignore the warnings issued by the tests on EEGLAB files
