### 2022-07-11 26a2f1f
Ignore RuntimeWarnings issued by the grid interpolation for ICLabel topographic feature (#69)
* ignore runtimewarnings during inteprolation

* remove ignore warnings from tests

* add to changelog

* fix tests for RuntimeWarning because of ICA method and instance BP filter

* fix style

* ignore the warnings issued by the tests on EEGLAB files
