+++ mne_icalabel/iclabel/tests/test_features.py
@@ -1,3 +1,4 @@
+import warnings
 from pathlib import Path
 
 import numpy as np
@@ -71,7 +72,6 @@ kwargs = {"raw": dict(preload=True), "epo": dict()}
 
 
 # ----------------------------------------------------------------------------
-@pytest.mark.filterwarnings("ignore::RuntimeWarning")
 @pytest.mark.parametrize(
     "file, psd_constant_file, eeglab_feature_file",
     [
@@ -87,7 +87,9 @@ def test_get_features_from_precomputed_ica(file, psd_constant_file, eeglab_featu
     ica = read_ica_eeglab(file)
 
     # Retrieve topo and autocorr
-    topo, _, autocorr = get_iclabel_features(inst, ica)
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore", category=RuntimeWarning)
+        topo, _, autocorr = get_iclabel_features(inst, ica)
 
     # Build PSD feature manually to match the subset
     # retrieve activation
@@ -147,7 +149,6 @@ def test_compute_ica_activations(file, eeglab_result_file):
 
 
 # ----------------------------------------------------------------------------
-@pytest.mark.filterwarnings("ignore::RuntimeWarning")
 @pytest.mark.parametrize(
     "file, eeglab_result_file",
     [(raw_eeglab_path, raw_topo1_path), (epo_eeglab_path, epo_topo1_path)],
@@ -172,7 +173,6 @@ def test_topoplotFast(file, eeglab_result_file):
     assert np.allclose(topo1, topo1_eeglab, equal_nan=True)
 
 
-@pytest.mark.filterwarnings("ignore::RuntimeWarning")
 @pytest.mark.parametrize(
     "file, eeglab_result_file",
     [
