+++ mne_icalabel/features/__init__.py
@@ -1,0 +1,1 @@
+"""Features for the ICLabel"""
