### 2022-07-18 30f176a
Update mne_icalabel/features/topomap.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
