### 2022-07-18 ed6d891
Update mne_icalabel/features/topomap.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
