### 2022-07-18 daad638
Update mne_icalabel/features/topomap.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
