### 2022-07-19 c151acc
Export to BIDS should work and then try to get time-serires plot
