+++ mne_icalabel/commands/__init__.py
@@ -1,1 +1,3 @@
 """Entry-points for mne-icalabel commands."""
+
+from .mne_gui_ic_annotation import main
