+++ mne_icalabel/__init__.py
@@ -7,5 +7,4 @@
 
 __version__ = "0.3dev0"
 
-from . import gui
 from .label_components import label_components  # noqa: F401
