### 2022-07-19 0b9ee44
topomap.py formatting and test for topomap generation
