+++ mne_icalabel/features/__init__.py
@@ -1,1 +1,2 @@
 """Features for the ICLabel"""
+from features import topomap
