### 2022-07-20 6dfdc80
fix timeSeries opening in new window
