### 2022-07-20 90d2f9e
Apply suggestions from code review
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
