### 2022-07-20 2517f3a
refactor and tests
