+++ mne_icalabel/gui/_label_components.py
@@ -1,112 +1,26 @@
-import platform
-from typing import Dict, List
+from typing import Dict, List, Tuple, Union
 
 from matplotlib import pyplot as plt
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
-from matplotlib.figure import Figure
+from mne import BaseEpochs
+from mne.io import BaseRaw
 from mne.preprocessing import ICA
+from mne.utils import _validate_type
 from mne.viz import set_browser_backend
 from qtpy.QtCore import Qt, Slot
 from qtpy.QtWidgets import (
     QAbstractItemView,
     QButtonGroup,
-    QGridLayout,
+    QLayout,
     QListWidget,
+    QGridLayout,
     QMainWindow,
     QPushButton,
     QVBoxLayout,
     QWidget,
-)
+    )
 
 from mne_icalabel.config import ICLABEL_LABELS_TO_MNE
-
-_CH_MENU_WIDTH = 30 if platform.system() == "Windows" else 10
-
-
-class TopomapFig(FigureCanvasQTAgg):
-    """Topographic map figure widget."""
-
-    def __init__(self, width=4, height=4, dpi=100):
-        self.fig = Figure(figsize=(width, height), dpi=dpi)
-        self.axes = self.fig.subplots()
-        self.fig.subplots_adjust(bottom=0, left=0, right=1, top=1, wspace=0, hspace=0)
-        super().__init__(self.fig)
-
-    def reset(self) -> None:
-        """Reset the topographic plot."""
-        self.axes.clear()
-
-    def redraw(self) -> None:
-        """Redraw the data."""
-        self.fig.canvas.draw()
-        self.fig.canvas.flush_events()
-
-
-class PowerSpectralDensityFig(FigureCanvasQTAgg):
-    """PSD figure widget."""
-
-    def __init__(self, width=4, height=4, dpi=100):
-        self.fig = Figure(figsize=(width, height), dpi=dpi)
-        self.axes = self.fig.subplots()
-        self.axes.axis("off")
-        super().__init__(self.fig)
-
-    def reset(self) -> None:
-        """Reset the PSD plot."""
-        self.axes.clear()
-
-    def redraw(self) -> None:
-        """Redraw the data."""
-        self.axes.set_title("")
-        self.fig.tight_layout()
-        self.fig.canvas.draw()
-        self.fig.canvas.flush_events()
-
-
-class TimeSeriesFig(FigureCanvasQTAgg):
-    """Dummy time-series figure widget."""
-
-    def __init__(self, width=4, height=4, dpi=100):
-        self.fig = Figure(figsize=(width, height), dpi=dpi)
-        self.axes = self.fig.subplots()
-        self.fig.subplots_adjust(bottom=0, left=0, right=1, top=1, wspace=0, hspace=0)
-        # clean up excess plot text, invert
-        self.axes.set_xticks([])
-        self.axes.set_yticks([])
-        super().__init__(self.fig)
-
-
-# TODO: Maybe that should inherit from a QGroupBox?
-class Labels(QWidget):
-    """Widget with labels as push buttons.
-
-    Only one of the labels can be selected at once.
-    """
-
-    def __init__(self, labels, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.buttonGroup = QButtonGroup()
-        self.buttonGroup.setExclusive(True)
-        layout = QVBoxLayout()
-        for k, label in enumerate(labels + ["Reset"]):
-            pushButton = Labels.create_pushButton(label)
-            self.buttonGroup.addButton(pushButton, k)
-            layout.addWidget(pushButton)
-        self.setLayout(layout)
-
-    @staticmethod
-    def create_pushButton(label):
[…]
