### 2022-07-20 0a68ae9
simpler handling of the central widget and its layout
