### 2022-07-20 e13048b
add type hint for labels2save
