+++ mne_icalabel/features/tests/test_topomap_feature.py
@@ -1,9 +1,9 @@
+import numpy as np
 from mne.datasets import testing
 from mne.io import read_raw
 from mne.preprocessing import ICA
-from numpy.typing import NDArray
 
-from mne_icalabel.features.topomap import get_topomaps
+from mne_icalabel.features import get_topomap, get_topomaps
 
 directory = testing.data_path() / "MEG" / "sample"
 raw = read_raw(directory / "sample_audvis_trunc_raw.fif", preload=False)
@@ -13,8 +13,14 @@ ica = ICA(n_components=5, method="picard")
 ica.fit(raw)
 
 
-def test_topomap_feature():
+def test_topomap_defaults():
     """Test scalp topography array generation"""
-    topo_array = get_topomaps(ica, picks="eeg")
-    assert isinstance(topo_array, NDArray)
-    assert topo_array.shape == (ica.n_components_, 64, 64)
+    topomaps = get_topomaps(ica, picks="eeg")
+    assert isinstance(topomaps, np.ndarray)
+    assert topomaps.shape == (ica.n_components_, 64, 64)
+    
+    # test single topo with fake data
+    data = np.random.randint(1, 10, len(raw.ch_names))
+    topomap = get_topomap(data, raw.info)
+    assert isinstance(topomap, np.ndarray)
+    assert topomap.shape == (64, 64)
