### 2022-07-21 29f9bb6
Update test_topomap_feature.py
