### 2022-07-21 2fff59e
Update test_topomap_feature.py
