+++ mne_icalabel/features/__init__.py
@@ -1,2 +1,3 @@
 """Features for the ICLabel"""
-from .topomap import get_topomap, get_topomaps
+
+from .topomap import get_topomap, get_topomaps  # noqa: F401
