### 2022-07-21 8cd00cf
check version of mne to support only 1.1.0 for topomap feature
