### 2022-07-21 6221679
more tests and expose the same arguments for both topomap functions
