### 2022-07-21 dd1a121
import mne docdict and retrieved needed keys
