### 2022-07-21 3bc476e
add test for docs.py
