+++ mne_icalabel/utils/_docs.py
@@ -14,12 +14,7 @@ from mne.utils.docs import docdict as docdict_mne
 docdict = {}
 
 # ---- Documentation to inc. from MNE ----
-keys = (
-    "border_topomap",
-    "extrapolate_topomap",
-    "outlines_topomap",
-    "picks_ica",
-)
+keys = ("verbose",)
 
 for key in keys:
     docdict[key] = docdict_mne[key]
@@ -133,8 +128,7 @@ def copy_doc(source: Callable) -> Callable:
     def wrapper(func):
         if source.__doc__ is None or len(source.__doc__) == 0:
             raise RuntimeError(
-                f"The docstring from {source.__name__} could not be copied "
-                "because it was empty."
+                f"The docstring from {source.__name__} could not be copied " "because it was empty."
             )
         doc = source.__doc__
         if func.__doc__ is not None:
