### 2022-07-21 9451c81
Fix error in unit test
