### 2022-07-21 01abe5a
corrected 'picks' for the integer use case
