### 2022-07-21 8f2e830
fix type-hint for docdict_indented and add test for fill_doc of function without docstring
