### 2022-07-21 33ca279
[MNT] Move utils to a separate module and add utilities for docstring and testing (#75)
* add misisng . between version and dev keyword

* move utils to separate module

* add requires_version

* import mne docdict and retrieved needed keys

* add decorators to fill and copy doc

* add test for docs.py

* fix tests

* add type-hints

* fix import order

* fix type hints

* fix type-hint for docdict_indented and add test for fill_doc of function without docstring
