### 2022-07-21 bb9720b
better handling of MNE version
