+++ doc/conf.py
@@ -89,6 +89,7 @@ numpydoc_xref_ignore = {
     'epochs', 'freqs', 'times', 'arrays', 'lists', 'func', 'n_nodes',
     'n_estimated_nodes', 'n_samples', 'n_channels', 'Renderer',
     'n_ytimes', 'n_ychannels', 'n_events', 'n_components', 'n_classes',
+    'n_pixels',
 }
 numpydoc_xref_aliases = {
     # Python
