+++ doc/whats_new.rst
@@ -22,7 +22,7 @@ Version 0.3 (Unreleased)
 ...
 
 Enhancements
-~~~~~~~~~~~~`
+~~~~~~~~~~~~
 
 -
 
