### 2022-07-21 317c959
get_topomap to get_topomap_array
