+++ mne_icalabel/features/__init__.py
@@ -1,3 +1,3 @@
 """Features for the ICLabel"""
 
-from .topomap import get_topomap, get_topomaps  # noqa: F401
+from .topomap import get_topomap_array, get_topomaps  # noqa: F401
