+++ doc/api.rst
@@ -55,7 +55,7 @@ Contains functions to extract features from `~mne.preprocessing.ICA` instance an
 .. autosummary::
    :toctree: generated/
 
-   get_topomap
+   get_topomap_array
    get_topomaps
 
 Annotating Components
