### 2022-07-21 da63959
Update mne_icalabel/features/topomap.py
