### 2022-07-21 a877171
replace docstring for outlines
