### 2022-07-21 239f8d3
[ENH] Add topographic map feature via MNE (#71)
* Topomap feature

Script to generate the topomap feature.

* Create __init__.py

* Rename topomap_feature.py to topomap.py

* Update mne_icalabel/features/topomap.py

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Update mne_icalabel/features/topomap.py

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Update mne_icalabel/features/topomap.py

Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>

* Update topomap.py

* Update topomap.py

* Update topomap.py

* Update topomap.py

* topom
[…]
