### 2022-07-22 6ba7fe9
Update whats_new.rst
