+++ doc/whats_new.rst
@@ -35,6 +35,7 @@ API
 ~~~
 
 - Add topographic feature using MNE with `~mne_icalabel.features.get_topomap_array` and `~mne_icalabel.features.get_topomaps` by `Anand Saini`_ and `Mathieu Scheltienne`_ (:gh:`71`)
+- Add psd feature using MNE with `~mne_icalabel.features.get_psds` by `Anand Saini`_ (:gh:`73`)
 
 Authors
 ~~~~~~~
