### 2022-07-22 dc0f87c
Update doc/whats_new.rst
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
