### 2022-07-22 e0dc282
Update whats_new.rst
