+++ .github/workflows/unit_tests.yml
@@ -156,7 +156,7 @@ jobs:
 
       - name: Upload coverage stats to codecov
         if: "matrix.os == 'ubuntu-latest'"
-        uses: codecov/codecov-action@v1
+        uses: codecov/codecov-action@v3
         with:
           file: ./coverage.xml
 
