### 2022-07-24 1cb507c
Fix example by closing figure
