+++ requirements_doc.txt
@@ -11,6 +11,3 @@ sphinxcontrib-bibtex
 mne-bids
 pooch
 matplotlib
-qtpy
-pyvista>=0.32
-pyvistaqt>=0.4
