### 2022-07-24 1e379c2
Apply suggestions from code review
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
