+++ mne_icalabel/annotation/bids.py
@@ -50,7 +50,7 @@ def write_components_tsv(ica: ICA, fname):
     # extract the component labels if they are present in the ICA instance
     if ica.labels_:
         for label, comps in ica.labels_.items():
-            this_status = "good" if label != "brain" else "bad"
+            this_status = "good" if label == "brain" else "bad"
             if label in ICLABEL_LABELS_TO_MNE:
                 for comp in comps:
                     status[comp] = this_status
