### 2022-07-24 928274b
Update mne_icalabel/commands/__init__.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
