+++ mne_icalabel/commands/__init__.py
@@ -1,2 +1,1 @@
 """Entry-points for mne-icalabel commands."""
-
