+++ mne_icalabel/utils/__init__.py
@@ -1,1 +1,2 @@
+from ._checks import _check_qt_version, _validate_inst_and_ica
 from ._docs import fill_doc
