+++ .circleci/config.yml
@@ -92,7 +92,10 @@ jobs:
               pip install --progress-bar off .
               pip install --upgrade --progress-bar off -r requirements_testing.txt
               pip install --upgrade --progress-bar off -r requirements_doc.txt
-
+              python -m pip uninstall -yq sphinx-gallery mne-qt-browser
+              # TODO: Revert to upstream/main once https://github.com/mne-tools/mne-qt-browser/pull/105 is merged
+              python -m pip install --upgrade --progress-bar off https://github.com/mne-tools/mne-qt-browser/zipball/main https://github.com/sphinx-gallery/sphinx-gallery/zipball/master
+        
         - save_cache:
             key: pip-cache
             paths:
@@ -103,6 +106,10 @@ jobs:
             paths:
               - ~/.local/lib/python3.8/site-packages
               - ~/.local/bin
+
+        - run:
+            name: Check Qt
+            command: LD_DEBUG=libs python -c "from PySide6.QtWidgets import QApplication, QWidget; app = QApplication([])"
 
         # Look at what we have and fail early if there is some library conflict
         - run:
