### 2022-07-25 46a116e
Get pyqt working
