+++ .circleci/config.yml
@@ -92,6 +92,7 @@ jobs:
               pip install --progress-bar off .
               pip install --upgrade --progress-bar off -r requirements_testing.txt
               pip install --upgrade --progress-bar off -r requirements_doc.txt
+              pip install --upgrade --progress-bar off PyQt5
               python -m pip uninstall -yq sphinx-gallery mne-qt-browser
               # TODO: Revert to upstream/main once https://github.com/mne-tools/mne-qt-browser/pull/105 is merged
               python -m pip install --upgrade --progress-bar off https://github.com/mne-tools/mne-qt-browser/zipball/main https://github.com/sphinx-gallery/sphinx-gallery/zipball/master
