+++ .github/workflows/unit_tests.yml
@@ -52,7 +52,8 @@ jobs:
       - name: Install dependencies
         run: |
           pip install --upgrade --progress-bar off pip setuptools wheel
-
+          pip install $STD_ARGS --only-binary ":all:" PyQt6 PyQt6-sip PyQt6-Qt6
+          
       # build with sdist directly
       - uses: actions/checkout@v3
       - name: Build sdist
