+++ .github/workflows/unit_tests.yml
@@ -119,9 +119,8 @@ jobs:
 
       - uses: actions/checkout@v3
 
-      - name: Install QT dependencies
-        run: |
-          sudo apt install -qq graphviz optipng libxft2 ffmpeg libsm6 libxext6
+      - run: ./tools/setup_circleci.sh
+        name: 'Setup xvfb'
 
       - name: Install mne-icalabel
         run: |
