+++ .github/workflows/unit_tests.yml
@@ -119,7 +119,7 @@ jobs:
 
       - uses: actions/checkout@v3
 
-      - run: ./tools/setup_circleci.sh
+      - run: ./scripts/setup_circleci.sh
         name: 'Setup xvfb'
 
       - name: Install mne-icalabel
