+++ .github/workflows/unit_tests.yml
@@ -119,8 +119,9 @@ jobs:
 
       - uses: actions/checkout@v3
 
-      - run: ./scripts/setup_circleci.sh
-        name: 'Setup xvfb'
+      - name: 'Setup xvfb'
+        if: "matrix.os == 'ubuntu-latest'"
+        run: ./scripts/setup_circleci.sh
 
       - name: Install mne-icalabel
         run: |
