+++ .circleci/config.yml
@@ -43,7 +43,7 @@ jobs:
             name: Set BASH_ENV
             command: |
               set -e
-              ./scripts/setup_circleci.sh
+              ./scripts/setup_xvfb.sh
               sudo apt install -qq graphviz optipng python3.8-venv python3-venv libxft2 ffmpeg
               python3.8 -m venv ~/python_env
               echo "set -e" >> $BASH_ENV
