+++ .github/workflows/unit_tests.yml
@@ -121,7 +121,7 @@ jobs:
 
       - name: 'Setup xvfb'
         if: "matrix.os == 'ubuntu-latest'"
-        run: ./scripts/setup_circleci.sh
+        run: ./scripts/setup_xvfb.sh
 
       - name: Install mne-icalabel
         run: |
