### 2022-07-26 b7e4860
make get_topomap_array private
