+++ mne_icalabel/features/topomap.py
@@ -62,7 +62,7 @@ def get_topomaps(
     topomaps = np.zeros((len(picks), res, res))
     for i, component in enumerate(picks):
         topo = np.flipud(
-            get_topomap_array(
+            _get_topomap_array(
                 data[component, :], ica.info, res, outlines, image_interp, border, extrapolate
             )
         )
@@ -74,7 +74,7 @@ def get_topomaps(
 
 
 @fill_doc
-def get_topomap_array(
+def _get_topomap_array(
     data: NDArray[float],
     pos: Info,
     res: int = 64,
