+++ mne_icalabel/features/tests/test_topomap.py
@@ -4,7 +4,7 @@ from mne.datasets import testing
 from mne.io import read_raw
 from mne.preprocessing import ICA
 
-from mne_icalabel.features import get_topomap_array, get_topomaps
+from mne_icalabel.features import _get_topomap_array, get_topomaps
 from mne_icalabel.utils._testing import requires_version
 
 directory = testing.data_path() / "MEG" / "sample"
@@ -24,7 +24,7 @@ def test_topomap_defaults():
 
     # test single topo with fake data
     data = np.random.randint(1, 10, len(raw.ch_names))
-    topomap = get_topomap_array(data, raw.info)
+    topomap = _get_topomap_array(data, raw.info)
     assert isinstance(topomap, np.ndarray)
     assert topomap.shape == (64, 64)
 
@@ -46,7 +46,7 @@ def test_topomap_arguments(picks, res):
     assert topomaps.shape == (n_components, res, res)
 
     data = np.random.randint(1, 10, len(raw.ch_names))
-    topomap = get_topomap_array(data, raw.info, res=res)
+    topomap = _get_topomap_array(data, raw.info, res=res)
     assert isinstance(topomap, np.ndarray)
     assert topomap.shape == (res, res)
 
