### 2022-07-29 2ddc063
[ENH] Draft GUI for labeling components (#66)
* Initial commit to show skeleton

* Some more components

* Add draft gui

* Adding the GUI with basic Python API

* Adding updated widget

* draft_2

* update logic

* some typos

* Adding testing script

* remove old version

* plot psd and topomap

* change dpi

* resize interface

* fix style

* add entry-point to run the GUI and app.exec() to start the Qt event loop

* fix style

* add saving of the selected labels to self.saved_labels and add a reset button

* Adding updated api TO SAVE components to ICA isntance.

* Adding test file for TOD
[…]
