+++ mne_icalabel/utils/_checks.py
@@ -1,9 +1,11 @@
+import sys
 from typing import Union
 
 from mne import BaseEpochs
+from mne.fixes import _compare_version
 from mne.io import BaseRaw
 from mne.preprocessing import ICA
-from mne.utils import _validate_type
+from mne.utils import _validate_type, warn
 
 
 def _validate_inst_and_ica(inst: Union[BaseRaw, BaseEpochs], ica: ICA):
@@ -16,3 +18,28 @@ def _validate_inst_and_ica(inst: Union[BaseRaw, BaseEpochs], ica: ICA):
             "The provided ICA instance was not fitted. Please use the '.fit()' method to "
             "determine the independent components before trying to label them."
         )
+
+
+def _check_qt_version(*, return_api=False):
+    """Check if Qt is installed."""
+    try:
+        from qtpy import API_NAME as api
+        from qtpy import QtCore
+    except Exception:
+        api = version = None
+    else:
+        try:  # pyside
+            version = QtCore.__version__
+        except AttributeError:
+            version = QtCore.QT_VERSION_STR
+        if sys.platform == "darwin" and api in ("PyQt5", "PySide2"):
+            if not _compare_version(version, ">=", "5.10"):
+                warn(
+                    f"macOS users should use {api} >= 5.10 for GUIs, "
+                    f"got {version}. Please upgrade e.g. with:\n\n"
+                    f'    pip install "{api}>=5.10"\n'
+                )
+    if return_api:
+        return version, api
+    else:
+        return version
