+++ setup.cfg
@@ -33,7 +33,7 @@ project_urls =
   Source = https://github.com/mne-tools/mne-icalabel
 
 [options]
-python_requires = ~= 3.7
+python_requires = >= 3.7
 install_requires =
   mne >= 1.0
   numpy >= 1.16.0
