### 2022-07-31 mne-tools/mne-icalabel#83 @TomDonoghue
Notes on JOSS paper
Some brief comments / suggestions on the current draft of the JOSS paper:
- lines 16-18: I find this sentence a little vague, and a couple sentences could clarify the details
    - how it makes this available: "this work *makes the popular Matlab-based ICLabel model available in MNE-Python*
    - something more on what "modern Pytorch format", as this is a bit unclear
- line 41: say "we will build" but since this project now exists, the tense shouldn't be future
- the relationship between Pytorch / Tensorflow, and what this project does related to them is unclear. For the M/EEG researcher who's looking into this project, I think it may be unclear how / why you jump between these different projects without detailing how they relate to each other and a more detailed description of the status of the old tool, and the update here.

Note: this issue is part of the JOSS review:
https://github.com/openjournals/joss-reviews/issues/4484
