### 2022-07-31 #4484 @TomDonoghue
Sorry for being slow on this - the past month ended up being busier than expected!

I have worked through the review / module check including getting the module installed locally, and running through the provided example on the documentation site. In general, I left some comments that are all on the original repository (in the issues linked above). They mostly detail some small comments on documentation and paper stuff (particularly the example), but they are all pretty minor things. Overall, I think the tool / module is well developed and organized, and I have no substantive comments on the code or project as a whole. Once the authors have a chance to respond to the minor points raised, I think the remaining things above can be checked off from my side.
