+++ setup.cfg
@@ -112,9 +112,15 @@ ignore-decorators = property
 match = ^(?!.*(setup|__init__|test_.*)).*\.py
 
 [coverage:run]
+branch = True
+source = mne_icalabel
 omit =
     # Do not include test script in coverage report
     *tests*
+    */bin/*
+    */setup.py
+    */mne_icalabel/conftest.py
+
 
 [coverage:report]
 # Regexes for lines to exclude from consideration
@@ -124,3 +130,6 @@ exclude_lines =
 
     # Don't complain if non-runnable code isn't run:
     if __name__ == .__main__.:
+
+    @abstractmethod
+    @abstractclassmethod
