### 2022-07-31 2bfbf4f
[MNT] - Small formatting updates (#85)
* fix formatting

* minor docstring updates

* Try to merge

* Clean up files

Co-authored-by: Adam Li <adam2392@gmail.com>
