+++ setup.cfg
@@ -54,11 +54,32 @@ universal = true
 
 # Testing options
 [flake8]
-exclude = __init__.py, mne_icalabel/**/_eeg_features.py
+exclude = 
+    __init__.py,mne_icalabel/**/_eeg_features.py
+    .venv
+    .git
+    __pycache__
+    doc/build
+    dist
+    .mypy_cache
+    doc/_build
+    doc/generated
+    doc/auto_examples
+
 ignore = W504,I101,I100,I201
+    # these rules don't play well with black
+    # whitespace before :
+    E203
+    # line break before binary operator
+    W503
+    E241,E305,W504,W605,E731
 per-file-ignores =
     mne_icalabel/tests/test_*.py:E402
+    # __init__.py files are allowed to have unused imports and lines-too-long
+    */__init__.py:F401
+    */**/**/__init__.py:F401,E501
 
+max-line-length = 115
 
 [check-manifest]
 ignore =
@@ -91,9 +112,15 @@ ignore-decorators = property
 match = ^(?!.*(setup|__init__|test_.*)).*\.py
 
 [coverage:run]
+branch = True
+source = mne_icalabel
 omit =
     # Do not include test script in coverage report
     *tests*
+    */bin/*
+    */setup.py
+    */mne_icalabel/conftest.py
+
 
 [coverage:report]
 # Regexes for lines to exclude from consideration
@@ -103,3 +130,6 @@ exclude_lines =
 
     # Don't complain if non-runnable code isn't run:
     if __name__ == .__main__.:
+
+    @abstractmethod
+    @abstractclassmethod
