### 2022-07-31 mne-tools/mne-icalabel#83 @adam2392
> Some brief comments / suggestions on the current draft of the JOSS paper:
>
> * lines 16-18: I find this sentence a little vague, and a couple sentences could clarify the details
>
>   * how it makes this available: "this work _makes the popular Matlab-based ICLabel model available in MNE-Python_
>   * something more on what "modern Pytorch format", as this is a bit unclear

Thanks for the comment, we've changed the sentence from: `This work adds
17 the popular ICLabel model (Pion-Tonachini et al., 2019) to the MNE-Python (Gramfort et al.,
18 2013) software toolkit in a modern Pytorch format (Paszke et al., 2019).` to now say the following:

```
This work makes the popular ICLabel model [@iclabel2019] available in Python by creating a software package compatible with MNE-Python [@Agramfort2013] software toolkit and a modern Pytorch format [@Pytorch2019]. The ICLabel model was previously only available in an outdated version of tensorflow that was no longer supported, and migrating the model now to an updated Pytorch version will ensure the model will not break due to unmaintained versions of software.
```

> * line 41: say "we will build" but since this project now exists, the tense shouldn't be future

Here, we meant to specify our future plans for the repository, which would extend beyond just the ICLabel model. I have modified the sentence to state:

```
In future versions, we plan on building on top of the ICLabel model to improve its robustness when operating on different types of recording hardware, sensor type, and sensor count.
```
