### 2022-07-31 mne-tools/mne-icalabel#83 @adam2392
> * the relationship between Pytorch / Tensorflow, and what this project does related to them is unclear. For the M/EEG researcher who's looking into this project, I think it may be unclear how / why you jump between these different projects without detailing how they relate to each other and a more detailed description of the status of the old tool, and the update here.

I added the following sentences in the `Statement of Need`.

`Pytorch is the most common Python deep-learning framework among researchers. Replicating the model in Pytorch enables future automatic ICA annotation research to easily build upon the model.`

Does this address this concern?
