+++ paper/paper.md
@@ -43,7 +43,7 @@ bibliography: paper.bib
 
 # Summary
 
-Scalp electroencephalography (EEG) and magnetoencephalography (MEG) analysis is typically very noisy and contains various non-neural signals, such as heart beat artifacts. Independent component analysis (ICA) is a common procedure to remove these artifacts [@Bell1995]. However, removing artifacts requires manual annotation of ICA components, which is subject to human error and very laborious when operating on large datasets. This work adds the popular ICLabel model [@iclabel2019] to the MNE-Python [@Agramfort2013] software toolkit in a modern Pytorch format [@Pytorch2019]. This enables the automatic labeling of ICA components, improving the preprocessing and analysis pipeline of electrophysiological data.
+Scalp electroencephalography (EEG) and magnetoencephalography (MEG) analysis is typically very noisy and contains various non-neural signals, such as heart beat artifacts. Independent component analysis (ICA) is a common procedure to remove these artifacts [@Bell1995]. However, removing artifacts requires manual annotation of ICA components, which is subject to human error and very laborious when operating on large datasets. This work makes the popular ICLabel model [@iclabel2019] available in Python by creating a software package compatible with MNE-Python [@Agramfort2013] software toolkit and a modern Pytorch format [@Pytorch2019]. The ICLabel model was previously only available in an outdated version of tensorflow that was no longer supported, and migrating the model now to an updated Pytorch version will ensure the model will not break due to unmaintained versions of software. This enables the automatic labeling of ICA components, improving the preprocessing and analysis pipeline of electrophysiological data.
 
 The Python ICLabel model is fully tested against and matches exactly the output produced in its MATLAB counterpart [@iclabel2019]. Moreover, this work builds the API on top of the robust MNE-Python ecosystem, enabling a seamless integration of automatic ICA analysis.
 
@@ -51,7 +51,9 @@ The Python ICLabel model is fully tested against and matches exactly the output 
 
 Typically EEG and MEG data have many artifacts due to a variety of signal sources, such as heartbeats, eye blinks, line noise, channel noise, or muscle artifacts. A common signal processing technique for assisting in separating true brain signals from noise sources is ICA [@Bell1995]. ICA performs blind-source separation decomposing the observed noisy electrophysiological signals into maximally independent sources. For example, it allows researchers to decompose EEG signals into brain signals and a heartbeat signal. Then by manually annotating a component time-series as a heartbeat, one can remove the heartbeat source from the signal and obtain their signals without the heartbeat. This process is typically manually carried out and requires a human to label each component that comes out from an ICA decomposition. This process is subject to human error and is difficult to scale up when dealing with high-dimensional EEG or MEG recordings.
 
-ICLabel was a proposed statistical model that uses neural networks and a crowdsourced training dataset to automatically label ICA components [@iclabel2019]. However, this package was previously only available in MATLAB, limiting its usage among Python neuroscience researchers. Moreover, the package relied on an outdated version of Tensorflow [@Tensorflow2016], which makes the model difficult to build on. In this work, we will build on top of the ICLabel model to improve its robustness when operating on different types of recording hardware, sensor type, and sensor count. Moreover, we plan on building new types of models that improve their overall accuracy and performance with respect to auto-labeling neural and non-neural signals. The availability of a simple API for ICLabel will facilitate a strong benchmark to build future models and bring in new developers and contributors.
[…]
