### 2022-07-31 mne-tools/mne-icalabel#83 @jacobf18
@adam2392 Small fix, but PyTorch is spelled with camel case.
