+++ README.md
@@ -71,8 +71,10 @@ Alternatively, you can also download a
 If you are interested in contributing, please read the
 [contributing guidelines](https://github.com/mne-tools/mne-icalabel/blob/main/CONTRIBUTING.md).
 
-# Forum
+# Getting Help
 
-Please visit the MNE forum to ask relevant questions.
+[<img alt="MNE Forum" src="https://user-images.githubusercontent.com/1681963/52239617-e2683480-289c-11e9-922b-5da55472e5b4.png" height=60/>](https://mne.discourse.group)
 
-https://mne.discourse.group
+For any usage questions, please post to the
+[MNE Forum](https://mne.discourse.group). Be sure to add the `mne-icalabel` tag to
+your question.
