### 2022-08-01 mne-tools/mne-icalabel#86 @agramfort
Done
