### 2022-08-01 mne-tools/mne-icalabel#86 @mscheltienne
Thank you for the detailed feedback! We will have a look at those items.
