### 2022-08-01 mne-tools/mne-icalabel#86 @adam2392
Thanks for the feedback! Will add a comment for each item.

> In the About part of the repo page (top right), it would be useful to put the documentation site in the link field

Added!

> There seems to be small issue with the version selector (multiple "v0.1"s)
> The License button in the header is a broken link

Fixed both license and drop-down menu: https://mne.tools/mne-icalabel/v0.1/index.html#indices-and-tables
