+++ requirements_doc.txt
@@ -1,13 +1,0 @@
-memory_profiler
-sphinx
-sphinx-gallery
-sphinx_rtd_theme
-sphinx-copybutton
-numpydoc
-pydata-sphinx-theme
-typing-extensions
-sphinx-autodoc-typehints
-sphinxcontrib-bibtex
-mne-bids
-pooch
-matplotlib
