+++ setup.cfg
@@ -1,86 +1,3 @@
-[metadata]
-name = mne-icalabel
-version = attr: mne_icalabel.__version__
-url = https://github.com/mne-tools/mne-icalabel
-author = mne-icalabel developers
-maintainer = Adam Li, Mathieu Scheltienne
-maintainer_email = ali39@jhu.edu
-description = MNE-ICALabel: Automatic labeling of ICA components from MEG, EEG and iEEG data with MNE.
-keywords = meg, eeg, ieeg, neuroscience
-long-description = file: README.md
-long-description-content-type = text/markdown; charset=UTF-8
-license = BSD-3-Clause
-license_files = LICENSE
-platforms = any
-classifiers =
-  Topic :: Scientific/Engineering
-  Intended Audience :: Science/Research
-  Intended Audience :: Developers
-  License :: OSI Approved
-  Topic :: Software Development
-  Topic :: Scientific/Engineering
-  Operating System :: Microsoft :: Windows
-  Operating System :: POSIX :: Linux
-  Operating System :: MacOS
-  Programming Language :: Python
-  Programming Language :: Python :: 3.7
-  Programming Language :: Python :: 3.8
-  Programming Language :: Python :: 3.9
-  Programming Language :: Python :: 3.10
-project_urls =
-  Documentation = https://mne.tools/mne-icalabel
-  Bug Reports = https://github.com/mne-tools/mne-icalabel/issues
-  Source = https://github.com/mne-tools/mne-icalabel
-
-[options]
-python_requires = ~= 3.7
-install_requires =
-  mne >= 1.0
-  numpy >= 1.16.0
-  scipy >= 1.2.0
-  torch
-  setuptools
-  importlib-resources; python_version<"3.9"
-packages = find:
-include_package_data = True
-
-[options.entry_points]
-console_scripts =
-    mne_gui_ic_annotation = mne_icalabel.commands.mne_gui_ic_annotation:main
-
-# Building package
-[bdist_wheel]
-universal = true
-
-# Testing options
-[flake8]
-exclude = 
-    __init__.py,mne_icalabel/**/_eeg_features.py
-    .venv
-    .git
-    __pycache__
-    doc/build
-    dist
-    .mypy_cache
-    doc/_build
-    doc/generated
-    doc/auto_examples
-
-ignore = W504,I101,I100,I201
-    # these rules don't play well with black
-    # whitespace before :
-    E203
-    # line break before binary operator
-    W503
-    E241,E305,W504,W605,E731
-per-file-ignores =
-    mne_icalabel/tests/test_*.py:E402
-    # __init__.py files are allowed to have unused imports and lines-too-long
-    */__init__.py:F401
-    */**/**/__init__.py:F401,E501
-
-max-line-length = 115
-
 [check-manifest]
 ignore =
     junit-results.xml
@@ -102,14 +19,6 @@ exclude = _script.py
 
 [mypy-tests.*]
 strict_optional = false
-
-[pydocstyle]
-convention = pep257
-match_dir = ^(?!\.|doc|examples).*$
-add-ignore = D100,D104,D107,D413
-add-select = D214,D215,D404,D405,D406,D407,D408,D409,D410,D411
-ignore-decorators = property
-match = ^(?!.*(setup|__init__|test_.*)).*\.py
 
 [coverage:run]
 branch = True
