+++ mne_icalabel/__init__.py
@@ -1,11 +1,5 @@
 """Automatic ICA labeling for MEG, EEG and iEEG data."""
 
-# Authors: Adam Li <ali39@jhu.edu>
-#          Jacob Feitelberg <>
-#
-# License: BSD (3-clause)
-
-__version__ = "0.3.dev0"
-
-from . import gui
+from . import gui  # noqa: F401
+from ._version import __version__  # noqa: F401
 from .label_components import label_components  # noqa: F401
