+++ mne_icalabel/_version.py
@@ -1,0 +1,5 @@
+"""Version number."""
+
+from importlib.metadata import version
+
+__version__ = version(__package__)
