### 2022-08-01 36f3c1f
include iclabel network weights
