+++ doc/conf.py
@@ -5,6 +5,7 @@ import warnings
 from datetime import date
 
 import mne
+from mne.fixes import _compare_version
 import sphinx_gallery  # noqa: F401
 from sphinx_gallery.sorting import ExampleTitleSortKey
 
@@ -237,6 +238,13 @@ else:
 
 os.environ['_MNE_BUILDING_DOC'] = 'true'
 scrapers = ('matplotlib',)
+try:
+    import mne_qt_browser
+    _min_ver = _compare_version(mne_qt_browser.__version__, '>=', '0.2')
+    if mne.viz.get_browser_backend() == 'qt' and _min_ver:
+        scrapers += (mne.viz._scraper._MNEQtBrowserScraper(),)
+except ImportError:
+    pass
 try:
     with warnings.catch_warnings():
         warnings.filterwarnings("ignore", category=DeprecationWarning)
