### 2022-08-01 45af3f2
add style check in separate workflow through github actions, remove all ignore words
