### 2022-08-01 57c51cc
remove configuration for mypy
