+++ setup.cfg
@@ -11,15 +11,6 @@ filterwarnings =
     # This is for Python 3.10+ and MNE <1.0
     ignore:The distutils package is deprecated.*:DeprecationWarning
 
-[mypy]
-ignore_missing_imports = true
-no_site_packages = true
-allow_redefinition = True
-exclude = _script.py
-
-[mypy-tests.*]
-strict_optional = false
-
 [coverage:run]
 branch = True
 source = mne_icalabel
