+++ pyproject.toml
@@ -109,7 +109,7 @@ test = [
 all = [
     'mne-icalabel[build]',
     'mne-icalabel[doc]',
-    'mne-icalabel[ica]'
+    'mne-icalabel[ica]',
     'mne-icalabel[style]',
     'mne-icalabel[test]',
 ]
