+++ mne_icalabel/_version.py
@@ -1,5 +1,8 @@
 """Version number."""
 
-from importlib.metadata import version
+try:
+    from importlib.metadata import version
+except ImportError:
+    from importlib_metadata import version
 
 __version__ = version(__package__)
