### 2022-08-01 faee59d
don't create file for a single word to ignore
