### 2022-08-01 73c321f
Revert "don't create file for a single word to ignore"
This reverts commit faee59d6feda58b7e30af44161ece859bc542446.
