+++ .github/workflows/code-style.yml
@@ -43,9 +43,6 @@ jobs:
         uses: psf/black@stable
         with:
           options: "--check --verbose"
-      - name: DEBUG
-        run: |
-          ls
       - name: Run codespell
         uses: codespell-project/actions-codespell@master
         with:
