### 2022-08-01 13775c6
match only mne_icalabel for pydocstyle
