### 2022-08-01 9ede40e
minimal mypy configuration
