+++ pyproject.toml
@@ -163,3 +163,7 @@ ignore-decorators = '(copy_doc|property|.*setter|.*getter|pyqtSlot|Slot)'
 match = '^(?!setup|__init__|test_).*\.py'
 match-dir = '^mne_icalabel.*'
 add_ignore = 'D100,D104,D107'
+
+[tool.mypy]
+ignore_missing_imports = true
+no_site_packages = true
