+++ doc/conf.py
@@ -1,5 +1,6 @@
 """Configure details for documentation with sphinx."""
 import os
+import subprocess
 import sys
 import warnings
 from datetime import date
@@ -202,8 +203,8 @@ html_sidebars = {
 
 html_context = {
     'versions_dropdown': {
-        'dev': 'v0.2 (devel)',
-        'stable': 'v0.1',
+        'dev': 'v0.3 (devel)',
+        'stable': 'v0.2',
         'v0.1': 'v0.1',
     },
 }
@@ -245,6 +246,7 @@ try:
         scrapers += (mne.viz._scraper._MNEQtBrowserScraper(),)
 except ImportError:
     pass
+
 try:
     with warnings.catch_warnings():
         warnings.filterwarnings("ignore", category=DeprecationWarning)
@@ -254,26 +256,45 @@ except Exception:
     pass
 else:
     scrapers += ('pyvista',)
+
 if 'pyvista' in scrapers:
     brain_scraper = mne.viz._brain._BrainScraper()
     scrapers = list(scrapers)
     scrapers.insert(scrapers.index('pyvista'), brain_scraper)
     scrapers = tuple(scrapers)
 
+compress_images = ('images', 'thumbnails')
+# let's make things easier on Windows users
+# (on Linux and macOS it's easy enough to require this)
+if sys.platform.startswith('win'):
+    try:
+        subprocess.check_call(['optipng', '--version'])
+    except Exception:
+        compress_images = ()
+
 sphinx_gallery_conf = {
-    'doc_module': 'mne_icalabel',
+    'doc_module': ('mne_icalabel',),
     'reference_url': {
         'mne_icalabel': None,
     },
-    'backreferences_dir': 'generated',
-    'plot_gallery': 'True',  # Avoid annoying Unicode/bool default warning
-    'within_subsection_order': ExampleTitleSortKey,
     'examples_dirs': ['../examples'],
     'gallery_dirs': ['auto_examples'],
-    'filename_pattern': '^((?!sgskip).)*$',
-    'matplotlib_animations': True,
-    'compress_images': ('images', 'thumbnails'),
+    'backreferences_dir': 'generated',
+    'plot_gallery': 'True',  # Avoid annoying Unicode/bool default warning
+    'thumbnail_size': (160, 112),
+    'remove_config_comments': True,
+    'min_reported_time': 1.,
+    'abort_on_example_error': False,
+    # 'reset_modules_order': 'both',
     'image_scrapers': scrapers,
+    'show_memory': not sys.platform.startswith(('win', 'darwin')),
+    'line_numbers': False,  # messes with style
+    'within_subsection_order': ExampleTitleSortKey,
+    'capture_repr': ('_repr_html_',),
+    'junit': os.path.join('..', 'test-results', 'sphinx-gallery', 'junit.xml'),
+    'matplotlib_animations': True,
+    'compress_images': compress_images,
+    'filename_pattern': '^((?!sgskip).)*$',
 }
 
 # sphinxcontrib-bibtex
