+++ mne_icalabel/iclabel/network.py
@@ -1,7 +1,7 @@
 try:
-    from importlib.resources import files
+    from importlib.resources import files  # type: ignore
 except ImportError:
-    from importlib_resources import files
+    from importlib_resources import files  # type: ignore
 
 import numpy as np
 import torch
