+++ mne_icalabel/_version.py
@@ -1,8 +1,8 @@
 """Version number."""
 
 try:
-    from importlib.metadata import version
+    from importlib.metadata import version  # type: ignore
 except ImportError:
-    from importlib_metadata import version
+    from importlib_metadata import version  # type: ignore
 
 __version__ = version(__package__)
