### 2022-08-01 10f1096
rework build test and run on windows as well
