### 2022-08-01 66c1cdb
move style back to main workflow file
