+++ .github/workflows/unit_tests.yml
@@ -13,6 +13,47 @@ on:
   workflow_dispatch:
 
 jobs:
+  style:
+    timeout-minutes: 10
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v3
+      - name: Setup Python 3.9
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.9'
+          architecture: 'x64'
+      - name: Install dependencies
+        run: |
+          python -m pip install --progress-bar off --upgrade pip setuptools wheel
+          python -m pip install --progress-bar off .[style]
+      - name: Setup flake8 annotations
+        uses: rbialon/flake8-annotations@v1
+      - name: Run flake8
+        uses: py-actions/flake8@v2
+        with:
+          path: "mne_icalabel"
+      - name: Run isort
+        uses: jamescurtin/isort-action@master
+      - name: Run black
+        uses: psf/black@stable
+        with:
+          options: "--check --verbose"
+      - name: Run codespell
+        uses: codespell-project/actions-codespell@master
+        with:
+          check_filenames: true
+          check_hidden: true
+          skip: './.git,./build'
+          ignore_words_file: ./.codespellignore
+      - name: Run pydocstyle
+        run: pydocstyle .
+      - name: Run mypy
+        # TODO: Should use the GitHub action when supports for pyproject.toml
+        # is added. https://github.com/jpetrucciani/mypy-check
+        run: mypy ./mne_icalabel
+
   build:
     strategy:
       fail-fast: false
