+++ .github/workflows/unit_tests.yml
@@ -152,7 +152,7 @@ jobs:
         run: |
           python -m pytest ./mne_icalabel --cov=mne_icalabel --cov-report=xml --cov-config=setup.cfg --verbose --ignore mne-python -vv mne_icalabel/gui
       - name: Upload coverage stats to codecov
-        if: "matrix.os == 'ubuntu-latest'"
+        if: "matrix.os == 'ubuntu'"
         uses: codecov/codecov-action@v3
         with:
           file: ./coverage.xml
