+++ .github/workflows/unit_tests.yml
@@ -64,6 +64,9 @@ jobs:
         python-version: [3.7, 3.8, 3.9, "3.10"]
     name: build ${{ matrix.os }} - py${{ matrix.python-version }}
     runs-on: ${{ matrix.os }}-latest
+    defaults:
+      run:
+        shell: bash
     steps:
       - name: Checkout repository
         uses: actions/checkout@v3
@@ -76,26 +79,24 @@ jobs:
         run: |
           python -m pip install --progress-bar off --upgrade pip setuptools wheel
           python -m pip install --progress-bar off .[build]
-      - name: Test and remove package install
-        run: |
-          python -c "import mne_icalabel; print(mne_icalabel.__version__)"
-          python -m pip uninstall -y mne_icalabel
+      - name: Test package install
+        run: python -c "import mne_icalabel; print(mne_icalabel.__version__)"
+      - name: Remove package install
+        run: python -m pip uninstall -y mne_icalabel
       - name: Build package
         run: python -m build
-      - name: Install, test and remove sdist
-        run: |
-          python -m pip install ./dist/*.tar.gz
-          python -c "import mne_icalabel; print(mne_icalabel.__version__)"
-          python -m pip uninstall -y mne_icalabel
-      - name: Install, test and remove wheel
-        run: |
-          python -m pip install ./dist/*.whl
-          python -c "import mne_icalabel; print(mne_icalabel.__version__)"
-          python -m pip uninstall -y mne_icalabel
-      - name: Test extras install
-        run: |
-          python -m pip install .[full]
-          python -c "import mne_icalabel; print(mne_icalabel.__version__)"
+      - name: Install sdist
+        run: pip install ./dist/*.tar.gz
+      - name: Test sdist install
+        run: python -c "import mne_icalabel; print(mne_icalabel.__version__)"
+      - name: Remove sdist install
+        run: python -m pip uninstall -y mne_icalabel
+      - name: Install wheel
+        run: pip install ./dist/*.whl
+      - name: Test wheel install
+        run: python -c "import mne_icalabel; print(mne_icalabel.__version__)"
+      - name: Remove wheel install
+        run: python -m pip uninstall -y mne_icalabel
 
   test:
     timeout-minutes: 30
