### 2022-08-02 0e4a0c4
check fmax, added tests
Check fmax value in psd.py.
added a few tests in tests_psd.py
