### 2022-08-02 adcd0f3
removed fmax from parametrize in test
