+++ .github/workflows/unit_tests.yml
@@ -149,10 +149,12 @@ jobs:
         run: |
           python -m pytest ./mne_icalabel --cov=mne_icalabel --cov-report=xml --cov-config=setup.cfg --verbose --ignore mne-python -vv mne_icalabel/gui
       - name: Upload coverage stats to codecov
-        if: "matrix.os == 'ubuntu'"
+        if: ${{ matrix.os == 'ubuntu' && matrix.python-version == "3.10" && matrix.mne == "stable" }}
         uses: codecov/codecov-action@v3
         with:
-          file: ./coverage.xml
+          files: ./coverage.xml
+          fail_ci_if_error: true # optional (default = false)
+          verbose: true # optional (default = false)
 
   release:
     name: Release
