### 2022-08-02 0d75479
fix require_version until 1.1 is released
