### 2022-08-02 8704918
simplify gui tests
