### 2022-08-02 e4db788
simplify ci for now
