+++ setup.cfg
@@ -1,7 +1,3 @@
-[check-manifest]
-ignore =
-    junit-results.xml
-
 [tool:pytest]
 addopts =
     --showlocals --durations=20 -ra --junit-xml=junit-results.xml
