+++ pyproject.toml
@@ -159,3 +159,24 @@ add_ignore = 'D100,D104,D107'
 [tool.mypy]
 ignore_missing_imports = true
 no_site_packages = true
+
+[tool.pytest]
+addopts = --verbose
+
+[tool.coverage.run]
+branch = true
+cover_pylib = false
+source = 'mne_icalabel'
+omit = [
+    'mne_icalabel/conftest.py',
+    'mne_icalabel/**/__init__.py',
+    'mne_icalabel/**/tests/**',
+    'mne_icalabel/commands/*',
+]
+
+[tool.coverage.report]
+exclude_lines = [
+    'pragma: no cover',
+    'if __name__ == .__main__.:',
+]
+precision = 2
