+++ setup.cfg
@@ -6,26 +6,3 @@ filterwarnings =
     error
     # This is for Python 3.10+ and MNE <1.0
     ignore:The distutils package is deprecated.*:DeprecationWarning
-
-[coverage:run]
-branch = True
-source = mne_icalabel
-omit =
-    # Do not include test script in coverage report
-    *tests*
-    */bin/*
-    */setup.py
-    */mne_icalabel/conftest.py
-
-
-[coverage:report]
-# Regexes for lines to exclude from consideration
-exclude_lines =
-    # Have to re-enable the standard pragma
-    pragma: no cover
-
-    # Don't complain if non-runnable code isn't run:
-    if __name__ == .__main__.:
-
-    @abstractmethod
-    @abstractclassmethod
