### 2022-08-02 dc82068
fix pytest config
