+++ .github/workflows/unit_tests.yml
@@ -149,7 +149,7 @@ jobs:
       - name: Download MNE testing dataset
         run: python -c "import mne; mne.datasets.testing.data_path()"
       - name: Run pytest
-        run: pytest main/mne_icalabel --cov=main/mne_icalabel --cov-report=xml --cov-config=pyproject.toml
+        run: pytest main/mne_icalabel --cov=main/mne_icalabel --cov-report=xml --cov-config=main/pyproject.toml
       - name: Upload coverage stats to codecov
         if: ${{ matrix.os == 'ubuntu' && matrix.python-version == '3.10' && matrix.mne == 'stable' }}
         uses: codecov/codecov-action@v3
