### 2022-08-02 66ec859
Update doc/index.rst
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
