### 2022-08-02 aa9020a
cache also mne-icalabel testing dataset
