+++ .github/workflows/unit_tests.yml
@@ -122,7 +122,7 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v3
         with:
-          path: main
+          path: main  # clone repository in a sub-directory
       - name: Setup Python ${{ matrix.python-version }}
         uses: actions/setup-python@v4
         with:
