+++ mne_icalabel/gui/tests/test_label_components.py
@@ -6,7 +6,6 @@ from mne.preprocessing import ICA
 from mne_icalabel.gui import label_ica_components
 from mne_icalabel.utils._testing import requires_version
 
-
 raw = read_raw_edf(testing.data_path() / "EDF" / "test_reduced.edf", preload=True)
 raw.filter(l_freq=1, h_freq=100)
 ica = ICA(n_components=15, random_state=12345)
