+++ mne_icalabel/__init__.py
@@ -1,4 +1,5 @@
 """Automatic ICA labeling for MEG, EEG and iEEG data."""
 
+from . import datasets  # noqa: F401
 from ._version import __version__  # noqa: F401
 from .label_components import label_components  # noqa: F401
