### 2022-08-02 870dd54
run mne-main only on ubuntu to reduce number of jobs
