### 2022-08-02 00a71ba
remove setup.cfg
