+++ setup.cfg
@@ -1,8 +1,0 @@
-[tool:pytest]
-addopts =
-    --showlocals --durations=20 -ra --junit-xml=junit-results.xml
-    --ignore=doc --ignore=examples
-filterwarnings =
-    error
-    # This is for Python 3.10+ and MNE <1.0
-    ignore:The distutils package is deprecated.*:DeprecationWarning
