### 2022-08-02 97f8b7d
simpler matrix
