### 2022-08-02 6ef88f6
retrieve and setup xvfb for ubuntu
