+++ .github/workflows/unit_tests.yml
@@ -121,6 +121,14 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           architecture: 'x64'
+      - name: Retrieve xvfb setup script
+        if: matrix.os == 'ubuntu'
+        run: |
+          curl https://raw.githubusercontent.com/mne-tools/mne-python/main/tools/setup_xvfb.sh -o setup_xvfb.sh
+      - name: Setup xvfb
+        if: matrix.os == 'ubuntu'
+        working-directory: main
+        run: ./setup_xvfb.sh
       - name: Install package
         run: |
           python -m pip install --progress-bar off --upgrade pip setuptools wheel
