+++ .github/workflows/unit_tests.yml
@@ -123,11 +123,9 @@ jobs:
           architecture: 'x64'
       - name: Retrieve xvfb setup script
         if: matrix.os == 'ubuntu'
-        run: |
-          curl https://raw.githubusercontent.com/mne-tools/mne-python/main/tools/setup_xvfb.sh -o setup_xvfb.sh
+        run: curl https://raw.githubusercontent.com/mne-tools/mne-python/main/tools/setup_xvfb.sh -o setup_xvfb.sh
       - name: Setup xvfb
         if: matrix.os == 'ubuntu'
-        working-directory: main
         run: ./setup_xvfb.sh
       - name: Install package
         run: |
