### 2022-08-03 f740a59
[DOC] Adding sections to paper (#87)
