+++ .circleci/config.yml
@@ -5,26 +5,6 @@ jobs:
       docker:
         - image: cimg/base:stable-20.04
       steps:
-        - restore_cache:
-            keys:
-              - source-cache-iclabel
-
-        - checkout
-
-        - run:
-            name: Complete checkout
-            command: |
-              if ! git remote -v | grep upstream; then
-                git remote add upstream https://github.com/mne-tools/mne-icalabel.git
-              fi
-              git remote set-url upstream https://github.com/mne-tools/mne-icalabel.git
-              git fetch upstream
-
-        - save_cache:
-            key: source-cache-iclabel
-            paths:
-              - ".git"
-
         - run:
             name: Check-skip
             command: |
@@ -34,16 +14,31 @@ jobs:
                 echo "Skip detected, exiting job ${CIRCLE_JOB} for PR ${CIRCLE_PULL_REQUEST}."
                 circleci-agent step halt;
               fi
-
+        - restore_cache:
+            keys:
+              - source-cache-iclabel
+        - run:
+            name: Complete checkout
+            command: |
+              if ! git remote -v | grep upstream; then
+                git remote add upstream https://github.com/mne-tools/mne-icalabel.git
+              fi
+              git remote set-url upstream https://github.com/mne-tools/mne-icalabel.git
+              git fetch upstream
+        - save_cache:
+            key: source-cache-iclabel
+            paths:
+              - ".git"
         - add_ssh_keys:
             fingerprints:
               - "d6:88:6b:a0:80:bf:14:8e:80:2e:ed:11:52:26:37:68"
-
         - run:
             name: Set BASH_ENV
             command: |
               set -e
-              ./scripts/setup_xvfb.sh
+              curl https://raw.githubusercontent.com/mne-tools/mne-python/main/tools/setup_xvfb.sh -o setup_xvfb.sh
+              chmod +x setup_xvfb.sh
+              ./setup_xvfb.sh
               sudo apt install -qq graphviz optipng python3.8-venv python3-venv libxft2 ffmpeg
               python3.8 -m venv ~/python_env
               echo "set -e" >> $BASH_ENV
@@ -57,11 +52,10 @@ jobs:
               echo "BASH_ENV:"
               cat $BASH_ENV
               mkdir -p ~/mne_icalabel_data
-              touch pattern.txt;
-
-        # make sure pandoc is on the system
-        - run: sudo apt update && sudo apt install -y pandoc optipng
-
+              touch pattern.txt
+        - run:
+            name: Setup pandoc
+            command: sudo apt update && sudo apt install -y pandoc optipng
         - run:
             name: Merge with upstream
             command: |
@@ -71,82 +65,59 @@ jobs:
                 echo "Merging $(cat merge.txt)";
                 git pull --ff-only upstream "refs/pull/$(cat merge.txt)/merge";
               fi
-
         - run:
             name: Install relevant dependencies
             command: |
               set -e
               ./scripts/circle_ci_dependencies.sh
-
-        # Load pip cache
         - restore_cache:
             keys:
               - pip-cache
         - restore_cache:
             keys:
               - user-install-bin-cachev1
-
         - run:
             name: Get Python running and install dependencies
             command: |
-              pip install --progress-bar off .
-              pip install --upgrade --progress-bar off -r requirements_testing.txt
-              pip install --upgrade --progress-bar off -r requirements_doc.txt
-              pip install --upgrade --progress-bar off PyQt5
-              python -m pip uninstall -yq sphinx-gallery mne-qt-browser
-              python -m pip install --upgrade --no-deps https://github.com/mne-tools/mne-python/archive/main.zip
[…]
