+++ .github/workflows/unit_tests.yml
@@ -121,14 +121,6 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           architecture: 'x64'
-      - name: Retrieve xvfb setup script
-        if: matrix.os == 'ubuntu'
-        run: |
-          curl https://raw.githubusercontent.com/mne-tools/mne-python/main/tools/setup_xvfb.sh -o setup_xvfb.sh
-          chmod +x setup_xvfb.sh
-      - name: Setup xvfb
-        if: matrix.os == 'ubuntu'
-        run: ./setup_xvfb.sh
       - name: Install package
         run: |
           python -m pip install --progress-bar off --upgrade pip setuptools wheel
@@ -153,8 +145,10 @@ jobs:
         run: |
           python -c "import mne; mne.datasets.testing.data_path()"
           python -c "import mne_icalabel; mne_icalabel.datasets.icalabel.data_path()"
-      - name: Run pytest
-        run: pytest main/mne_icalabel --cov=main/mne_icalabel --cov-report=xml --cov-config=main/pyproject.toml
+      - name: Run pytest  # headless via Xvfb on linux
+        uses: GabrielBB/xvfb-action@v1.6
+        with:
+          run: pytest main/mne_icalabel --cov=main/mne_icalabel --cov-report=xml --cov-config=main/pyproject.toml
       - name: Upload coverage stats to codecov
         if: ${{ matrix.os == 'ubuntu' && matrix.python-version == '3.10' && matrix.mne == 'stable' }}
         uses: codecov/codecov-action@v3
