### 2022-08-03 94a80be
try to install qt5 dep
