+++ .github/workflows/unit_tests.yml
@@ -121,6 +121,8 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
           architecture: 'x64'
+      - name: Install Qt dependencies
+        run: apt-get install qt5-default
       - name: Install package
         run: |
           python -m pip install --progress-bar off --upgrade pip setuptools wheel
