+++ .github/workflows/unit_tests.yml
@@ -122,7 +122,10 @@ jobs:
           python-version: ${{ matrix.python-version }}
           architecture: 'x64'
       - name: Install Qt dependencies
-        run: apt-get install qt5-default
+        if: "matrix.os == 'ubuntu'"
+        run: |
+          sudo apt update
+          sudo apt-get install qt5-default
       - name: Install package
         run: |
           python -m pip install --progress-bar off --upgrade pip setuptools wheel
