### 2022-08-03 6c8c3f6
cleaning one by one: pytest-qt
