+++ pyproject.toml
@@ -95,7 +95,6 @@ test = [
     'pytest>=6.0.0',
     'pytest-cov',
     'pytest-harvest',
-    'pytest-qt',
     'pytest-timeout',
 ]
 all = [
