### 2022-08-03 13931b8
cleaning one-by-one: pydocstyle, pymatreader
