+++ pyproject.toml
@@ -40,6 +40,7 @@ dependencies = [
     'numpy >= 1.16.0',
     'scipy >= 1.2.0',
     'mne >= 1.0',
+    'pooch',
     'torch',
     'importlib-metadata; python_version<"3.8"',
     'importlib-resources; python_version<"3.9"',
@@ -89,7 +90,6 @@ test = [
     'matplotlib',
     'mne-bids',
     'pandas',
-    'pooch',
     'pymatreader',
     'pytest',
     'pytest-cov',
