### 2022-08-03 dcac6ac
move matplotlib to gui dep.
