### 2022-08-03 6a132d5
revert, Xvfb is needed as it should be
