+++ .github/workflows/unit_tests.yml
@@ -135,18 +135,18 @@ jobs:
         run: |
           python -m pip uninstall -y mne
           python -m pip install git+https://github.com/mne-tools/mne-python
-      - name: Display MNE infos
-        run: mne sys_info
-      - name: Retrieve MNE/ICA-Label testing dataset version
-        run: |
-          curl https://raw.githubusercontent.com/mne-tools/mne-testing-data/master/version.txt -o mne_testing_data_version.txt
-          curl https://raw.githubusercontent.com/adam2392/mne-testing-icalabel-data/main/version.txt -o mne_icalabel_testing_data_version.txt
-      - name: Cache MNE/ICA-Label testing dataset
-        uses: actions/cache@v3
-        with:
-          key: ${{ matrix.os }}-${{ hashFiles('mne_testing_data_version.txt') }}-${{ hashFiles('mne_icalabel_testing_data_version.txt') }}
-          path: ~/mne_data
-      - name: Download MNE/ICA-Label testing dataset
+      # - name: Display MNE infos
+      #   run: mne sys_info
+      #  - name: Retrieve MNE/ICA-Label testing dataset version
+      #    run: |
+      #      curl https://raw.githubusercontent.com/mne-tools/mne-testing-data/master/version.txt -o mne_testing_data_version.txt
+      #      curl https://raw.githubusercontent.com/adam2392/mne-testing-icalabel-data/main/version.txt -o mne_icalabel_testing_data_version.txt
+      #  - name: Cache MNE/ICA-Label testing dataset
+      #    uses: actions/cache@v3
+      #    with:
+      #      key: ${{ matrix.os }}-${{ hashFiles('mne_testing_data_version.txt') }}-${{ hashFiles('mne_icalabel_testing_data_version.txt') }}
+      #      path: ~/mne_data
+       - name: Download MNE/ICA-Label testing dataset
         run: |
           python -c "import mne; mne.datasets.testing.data_path()"
           python -c "import mne_icalabel; mne_icalabel.datasets.icalabel.data_path()"
