### 2022-08-03 cebfcb7
Revert "try without conftest.py"
This reverts commit cd5f50265921f7cb0844165bb9f5868eba37f10d.
