### 2022-08-03 75c6352
try without unused matplotlib fixtures
