### 2022-08-03 37c9fd0
try with empty pytest_configure
