+++ mne_icalabel/conftest.py
@@ -1,8 +1,0 @@
-# Author: Eric Larson <larson.eric.d@gmail.com>
-#
-# License: BSD-3-Clause
-
-
-def pytest_configure(config):
-    """Configure pytest options."""
-    pass
