+++ .github/workflows/unit_tests.yml
@@ -144,7 +144,7 @@ jobs:
       - name: Cache MNE/ICA-Label testing dataset
         uses: actions/cache@v3
         with:
-          key: ${{ matrix.os }}-${{ hashFiles('mne_testing_data_version.txt') }}-${{ hashFiles('mne_icalabel_testing_data_version.txt') }}
+          key: ${{ matrix.os }}-${{ matrix.mne }}-${{ hashFiles('mne_testing_data_version.txt') }}-${{ hashFiles('mne_icalabel_testing_data_version.txt') }}
           path: ~/mne_data
       - name: Download MNE/ICA-Label testing dataset
         run: |
