### 2022-08-03 mne-tools/mne-icalabel#86 @adam2392
Hi @TomDonoghue I have addressed the remainder of your comments around the example and documentation: https://mne.tools/mne-icalabel/dev/index.html

Thanks so much for the feedback! lmk if you have other issues w/ how we reformatted/re-wrote the example.
