### 2022-08-03 #4484 @adam2392
Hi @TomDonoghue I have addressed your points raised in the https://github.com/mne-tools/mne-icalabel/issues/86#issuecomment-1204146456 issue. Thanks!
