+++ .circleci/config.yml
@@ -88,8 +88,9 @@ jobs:
               python -m pip install --progress-bar off --upgrade pip setuptools wheel
               python -m pip install --progress-bar off .[doc,gui]
               python -m pip uninstall mne mne-qt-browser
-              python -m pip install git+https://github.com/mne-tools/mne-python
-              python -m pip install git+https://github.com/mne-tools/mne-qt-browser
+              python -m pip install --progress-bar off git+https://github.com/mne-tools/mne-python
+              python -m pip install --progress-bar off git+https://github.com/mne-tools/mne-qt-browser
+
         - save_cache:
             key: pip-cache
             paths:
