+++ .circleci/config.yml
@@ -87,7 +87,7 @@ jobs:
             command: |
               python -m pip install --progress-bar off --upgrade pip setuptools wheel
               python -m pip install --progress-bar off .[doc,gui]
-              python -m pip uninstall mne mne-qt-browser
+              python -m pip uninstall -yq mne mne-qt-browser
               python -m pip install --progress-bar off git+https://github.com/mne-tools/mne-python
               python -m pip install --progress-bar off git+https://github.com/mne-tools/mne-qt-browser
 
