### 2022-08-03 103f728
bump mne req to 1.1
