+++ pyproject.toml
@@ -39,7 +39,7 @@ classifiers = [
 dependencies = [
     'numpy >= 1.16.0',
     'scipy >= 1.2.0',
-    'mne >= 1.0',
+    'mne >= 1.1',
     'pooch',
     'torch',
     'importlib-metadata; python_version<"3.8"',
