### 2022-08-03 bbc287c
use qt6 in CIs
