### 2022-08-03 011c7d1
use -yq flag to uninstall and suppress verbosity
