### 2022-08-03 1f7fb39
remove requires_version (fixed in mne 1.1)
