### 2022-08-03 0dbb038
remove requires_version from gui tests
