### 2022-08-03 e7504a2
try with python3-pyqt6
