### 2022-08-03 13e3635
revert to qt5 for tests
