+++ pyproject.toml
@@ -69,7 +69,6 @@ doc = [
 gui = [
     'matplotlib',
     'mne-qt-browser',
-    'PyQt6',
     'qtpy',
 ]
 ica = [
@@ -90,6 +89,7 @@ test = [
     'mne-icalabel[ica]',
     'mne-bids',
     'pandas',
+    'PyQt5',
     'pymatreader',
     'pytest',
     'pytest-cov',
