### 2022-08-03 33b9cc2
try apt-get
