+++ .github/workflows/unit_tests.yml
@@ -124,8 +124,8 @@ jobs:
       - name: Install Qt dependencies
         if: "matrix.os == 'ubuntu'"
         run: |
-          sudo apt update
-          sudo apt install qt5-defaults
+          sudo apt-get update
+          sudo apt-get install qt5-defaults
       - name: Install package
         run: |
           python -m pip install --progress-bar off --upgrade pip setuptools wheel
