### 2022-08-03 #4484 @emdupre
Thanks, @TomDonoghue and @adam2392 !

> Generally what sort of timeline should we expect?

We ask that all reviews be completed [within six weeks](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) (so from our initial start date of 18 June, I wouldn't expect reviews before the end of last week). Given ongoing summer holidays, though, many folks are in-and-out of office so I'm generally assuming about a two-week buffer period.

Just as an explicit update : I'm following up with @hvgazula via email in case he's missing these GitHub notifications. And thank you again, @TomDonoghue, for your review on this !
