+++ .circleci/config.yml
@@ -86,7 +86,7 @@ jobs:
             name: Get Python running and install dependencies
             command: |
               python -m pip install --progress-bar off --upgrade pip setuptools wheel
-              python -m pip install --progress-bar off .
+              python -m pip install --progress-bar off -e .
               python -m pip install --progress-bar off .[doc,gui]
               python -m pip install --progress-bar off git+https://github.com/mne-tools/mne-python
               python -m pip install --progress-bar off git+https://github.com/mne-tools/mne-qt-browser
