+++ .circleci/config.yml
@@ -51,6 +51,7 @@ jobs:
               chmod +x setup_xvfb.sh
               ./setup_xvfb.sh
               sudo apt install -qq graphviz optipng python3.8-venv python3-venv libxft2 ffmpeg
+              sudo apt install qt5-default
               python3.8 -m venv ~/python_env
               echo "set -e" >> $BASH_ENV
               echo "export OPENBLAS_NUM_THREADS=4" >> $BASH_ENV
@@ -70,14 +71,6 @@ jobs:
         - run:
             name: DEBUG
             command: ls -l
-        # - run:
-        #     name: Install relevant dependencies
-        #     command: |
-        #       set -e
-        #       curl https://raw.githubusercontent.com/mne-tools/mne-python/main/tools/circle_ci_dependencies.sh -o circle_ci_dependencies.sh
-        #       chmod +x circle_ci_dependencies.sh
-        #       ./circle_ci_dependencies.sh
-
         - restore_cache:
             keys:
               - pip-cache
