+++ pyproject.toml
@@ -71,6 +71,7 @@ gui = [
     'matplotlib',
     'mne-qt-browser',
     'qtpy',
+    'PySide6!=6.3.0',
 ]
 ica = [
     'joblib',
