### 2022-08-04 e171e47
replace scikit-learn doc dep. with ICA key
