+++ .circleci/config.yml
@@ -8,7 +8,6 @@ jobs:
         - restore_cache:
             keys:
               - source-cache-iclabel
-        # needed
         - checkout
         - run:
             name: Complete checkout
@@ -83,10 +82,9 @@ jobs:
               python -m pip install --progress-bar off --upgrade pip setuptools wheel
               python -m pip install --progress-bar off .
               python -m pip install --progress-bar off .[doc,gui]
+              python -m pip uninstall -yq mne mne-qt-browser
               python -m pip install --progress-bar off git+https://github.com/mne-tools/mne-python
               python -m pip install --progress-bar off git+https://github.com/mne-tools/mne-qt-browser
-              # python -m pip uninstall -yq mne mne-qt-browser
-
         - save_cache:
             key: pip-cache
             paths:
