+++ .circleci/config.yml
@@ -67,9 +67,6 @@ jobs:
         - run:
             name: Setup pandoc
             command: sudo apt update && sudo apt install -y pandoc optipng
-        - run:
-            name: DEBUG
-            command: ls -l
         - restore_cache:
             keys:
               - pip-cache
