+++ pyproject.toml
@@ -174,7 +174,7 @@ source = [
     'mne_icalabel',
 ]
 omit = [
-    'mne_icalabel/**/__init__.py',
+    '**/__init__.py',
     'mne_icalabel/commands/*',
     '**/tests/**',
 ]
