+++ pyproject.toml
@@ -59,6 +59,7 @@ doc = [
     'numpydoc',
     'pooch',
     'pydata-sphinx-theme',
+    'PyQt5',
     'sphinx',
     'sphinxcontrib-bibtex',
     'sphinx-autodoc-typehints',
@@ -71,7 +72,6 @@ gui = [
     'matplotlib',
     'mne-qt-browser',
     'qtpy',
-    'PySide6!=6.3.0',
 ]
 ica = [
     'joblib',
