+++ .circleci/config.yml
@@ -92,9 +92,6 @@ jobs:
               - ~/.local/lib/python3.8/site-packages
               - ~/.local/bin
         - run:
-            name: Check Qt
-            command: LD_DEBUG=libs python -c "from PySide6.QtWidgets import QApplication, QWidget; app = QApplication([])"
-        - run:
             name: Check installation
             command: |
                which python
