### 2022-08-04 1b633e0
remove set -e
