### 2022-08-04 65f8208
split debug information and add caching of the MNE sample dataset
