+++ .circleci/config.yml
@@ -91,16 +91,28 @@ jobs:
               - ~/.local/lib/python3.8/site-packages
               - ~/.local/bin
         - run:
+            name: Check pip package versions
+            command: pip freeze
+        - run:
+            name: Display MNE infos
+            command: QT_DEBUG_PLUGINS=1 mne sys_info -pd
+        - run:
             name: Check installation
             command: |
-               which python
-               pip freeze
-               python -c "print('Imported mne-icalabel correctly'); import mne_icalabel;"
-               QT_DEBUG_PLUGINS=1 mne sys_info -pd
                LIBGL_DEBUG=verbose python -c "import matplotlib.pyplot as plt; plt.figure()"
                python -c "import mne; mne.set_config('MNE_LOGGING_LEVEL', 'info')"
                python -c "import mne; level = mne.get_config('MNE_LOGGING_LEVEL'); assert level.lower() == 'info', repr(level)"
                python -c "import numpy; numpy.show_config()"
+        - restore_cache:
+            keys:
+              - mne-data
+        - run:
+            name: Download MNE sample dataset
+            command: python -c "import mne; mne.datasets.sample.data_path()"
+        - save_cache:
+            key: mne-data
+            paths:
+              - ~/mne_data
         - run:
             name: Build documentation
             command: |
