### 2022-08-04 1a0093e
revert caching of the sample datasets as examples are not always rendered
