+++ .circleci/config.yml
@@ -103,16 +103,6 @@ jobs:
                python -c "import mne; mne.set_config('MNE_LOGGING_LEVEL', 'info')"
                python -c "import mne; level = mne.get_config('MNE_LOGGING_LEVEL'); assert level.lower() == 'info', repr(level)"
                python -c "import numpy; numpy.show_config()"
-        - restore_cache:
-            keys:
-              - mne-data
-        - run:
-            name: Download MNE sample dataset
-            command: python -c "import mne; mne.datasets.sample.data_path()"
-        - save_cache:
-            key: mne-data
-            paths:
-              - ~/mne_data
         - run:
             name: Build documentation
             command: |
