### 2022-08-04 078e5cc
does pattern.txt exists?
