### 2022-08-04 5b9f152
remove setup of xvfb
