+++ .circleci/config.yml
@@ -45,9 +45,6 @@ jobs:
         - run:
             name: Set BASH_ENV
             command: |
-              curl https://raw.githubusercontent.com/mne-tools/mne-python/main/tools/setup_xvfb.sh -o setup_xvfb.sh
-              chmod +x setup_xvfb.sh
-              ./setup_xvfb.sh
               sudo apt install -qq graphviz optipng python3.8-venv python3-venv libxft2 ffmpeg
               sudo apt install qt5-default
               python3.8 -m venv ~/python_env
