### 2022-08-04 601e286
Revert "remove setup of xvfb"
This reverts commit 5b9f152075e635f593d4c610b047d97ed3b7ddaa.
