+++ .circleci/config.yml
@@ -49,7 +49,6 @@ jobs:
               chmod +x setup_xvfb.sh
               ./setup_xvfb.sh
               sudo apt install -qq graphviz optipng python3.8-venv python3-venv libxft2 ffmpeg
-              sudo apt install qt5-default
               python3.8 -m venv ~/python_env
               echo "set -e" >> $BASH_ENV
               echo "export OPENBLAS_NUM_THREADS=4" >> $BASH_ENV
