### 2022-08-04 c150fee
add back set -e to exit upon error and remove mne_icalabel_data directory
