### 2022-08-04 a9fd80d
add ls to debug
