### 2022-08-04 0995824
Revert "does circle support xvfb natively?"
This reverts commit a9bfc90fe38ab89fb95ab65039cc6f2add4c6d7b.
