### 2022-08-04 cc372ad
Revert "remove unused memory_profiler"
This reverts commit 1d574fb0d3ec3e99aad3689e4ceab3e63706f763.
