+++ pyproject.toml
@@ -54,6 +54,7 @@ build = [
 doc = [
     'mne-icalabel[gui]',
     'mne-icalabel[ica]',
+    'memory_profiler',
     'numpydoc',
     'pydata-sphinx-theme',
     'PyQt5',
