### 2022-08-04 ffd29de
sidecar .json file saved separately
Used suffix 'digitizer' to save it separately  from the components label .tsv file in mne_iclabel/annotations/mara.py
