+++ .circleci/config.yml
@@ -249,13 +249,13 @@ workflows:
   
   main:
     jobs:
-      - build_docs:
+      - build_doc:
           scheduled: "true"
-          name: build_docs_main
+          name: build_doc_main
       - deploy:
           name: deploy_main
           requires:
-            - build_docs_main
+            - build_doc_main
     triggers:
       - schedule:
           # "At 00:00" (once a day) should be enough "0 0 * * *",
