+++ .circleci/config.yml
@@ -253,7 +253,7 @@ workflows:
           scheduled: "true"
           name: build_doc_main
       - deploy:
-          name: doc_deploy_main
+          name: doc-deploy_main
           requires:
             - build_doc_main
     triggers:
