+++ .circleci/config.yml
@@ -143,7 +143,7 @@ jobs:
           type: string
           default: "false"
       docker:
-        - image: circleci/python:3.9.2-buster
+        - image: cimg/python:3.9.2-buster
       steps:
         - restore_cache:
             keys:
