### 2022-08-04 f964898
upload artifacts right after building
