### 2022-08-05 45ff734
add additional skip folders to codespell action
