### 2022-08-05 a5cdd8f
[MNT] Improve packaging, managing of dependencies, and managing of ignored warning/errors (#88)
* migrate to pyproject.toml

* handle version

* drop manifest since tarball does not need any files outside mne_icalabel

* include iclabel network weights

* simpler setup.py

* run draft-pdf only on push with changes to the paper

* add style check in separate workflow through github actions, remove all ignore words

* remove configuration for mypy

* fix typos

* fix typo

* fix flake8

* fix codespell

* don't create file for a single word to ignore

* Revert "don't create file for a single wor
[…]
