+++ mne_icalabel/_version.py
@@ -1,0 +1,8 @@
+"""Version number."""
+
+try:
+    from importlib.metadata import version  # type: ignore
+except ImportError:
+    from importlib_metadata import version  # type: ignore
+
+__version__ = version(__package__)
