+++ mne_icalabel/utils/_checks.py
@@ -1,5 +1,5 @@
 import sys
-from typing import Union
+from typing import Tuple, Union
 
 from mne import BaseEpochs
 from mne.fixes import _compare_version
@@ -20,12 +20,27 @@ def _validate_inst_and_ica(inst: Union[BaseRaw, BaseEpochs], ica: ICA):
         )
 
 
-def _check_qt_version(*, return_api=False):
-    """Check if Qt is installed."""
+def _check_qt_version(raise_on_error: bool = False) -> Union[Tuple[None, None], Tuple[str, str]]:
+    """Check if Qt is available.
+
+    Parameters
+    ----------
+    raise_on_error : bool
+        If True, missing Qt bindings will raise an error. Else 'api' and 'version' are set to None.
+
+    Returns
+    -------
+    api : str | None
+        Which API is used. One of 'PyQt5', 'PyQt6', 'PySide2', 'PySide6'.
+    version : str | None
+        Version of the API.
+    """
     try:
         from qtpy import API_NAME as api
         from qtpy import QtCore
     except Exception:
+        if raise_on_error:
+            raise
         api = version = None
     else:
         try:  # pyside
@@ -39,7 +54,4 @@ def _check_qt_version(*, return_api=False):
                     f"got {version}. Please upgrade e.g. with:\n\n"
                     f'    pip install "{api}>=5.10"\n'
                 )
-    if return_api:
-        return version, api
-    else:
-        return version
+    return api, version
