+++ mne_icalabel/gui/_label_components.py
@@ -29,6 +29,7 @@ class ICAComponentLabeler(QMainWindow):
     ----------
     inst : Raw | Epochs
     ica : ICA
+    show : bool
     """
 
     def __init__(self, inst: Union[BaseRaw, BaseEpochs], ica: ICA, show: bool = True) -> None:
@@ -162,8 +163,8 @@ class ICAComponentLabeler(QMainWindow):
         _validate_type(ica, ICA, "ica", "ICA")
         if ica.current_fit == "unfitted":
             raise ValueError(
-                "ICA instance should be fit on the raw data before "
-                "running the ICA labeling GUI. Run `ica.fit(inst)`."
+                "ICA instance should be fitted on raw/epochs data before "
+                "running the ICA labeling GUI. Run 'ica.fit(inst)'."
             )
 
     # - Properties ------------------------------------------------------------
