+++ examples/iclabel_automatic_artifact_correction_ica.py
@@ -10,7 +10,9 @@ the ICLabel model :footcite:`iclabel2019`, which originates in EEGLab.
 For conceptual background on ICA, see :ref:`this scikit-learn tutorial
 <sphx_glr_auto_examples_decomposition_plot_ica_blind_source_separation.py>`.
 For a basic understanding of how to use ICA to remove artifacts, see `the
-tutorial <https://mne.tools/stable/auto_tutorials/preprocessing/40_artifact_correction_ica.html>`_ in MNE-Python.
+tutorial
+<https://mne.tools/stable/auto_tutorials/preprocessing/40_artifact_correction_ica.html>`_
+in MNE-Python.
 
 We begin as always by importing the necessary Python modules and loading some
 :ref:`example data <sample-dataset>`. Because ICA can be computationally
