+++ mne_icalabel/datasets/__init__.py
@@ -1,0 +1,3 @@
+from . import icalabel
+
+__all__ = ["icalabel"]

[…]
