+++ .circleci/config.yml
@@ -159,6 +159,12 @@ jobs:
               set -e
               echo "set -e" >> $BASH_ENV
               echo "export PATH=~/.local/bin/:$PATH" >> $BASH_ENV
+              python3.8 -m venv ~/python_env
+              echo "source ~/python_env/bin/activate" >> $BASH_ENV
+              mkdir -p ~/.local/bin
+              ln -s ~/python_env/bin/python ~/.local/bin/python
+              echo "BASH_ENV:"
+              cat $BASH_ENV
         - restore_cache:
             keys:
               - pip-cache
