+++ doc/api.rst
@@ -55,8 +55,8 @@ Contains functions to extract features from `~mne.preprocessing.ICA` instance an
 .. autosummary::
    :toctree: generated/
 
-   get_topomaps
    get_psds
+   get_topomaps
 
 Annotating Components
 =====================
