+++ mne_icalabel/features/__init__.py
@@ -1,4 +1,6 @@
 """Features for the ICLabel"""
 
-from .psd import get_psds  # noqa: F401
-from .topomap import get_topomaps  # noqa: F401
+from .psd import get_psds
+from .topomap import get_topomaps
+
+__all__ = ("get_psds", "get_topomaps")
