+++ mne_icalabel/features/psd.py
@@ -19,8 +19,8 @@ def get_psds(
     picks=None,
     reject="auto",
     reject_by_annotation: bool = False,
-    fmin: Optional[float] = 0,
-    fmax: Optional[float] = np.inf,
+    fmin: float = 0,
+    fmax: float = np.inf,
     tmin: Optional[float] = None,
     tmax: Optional[float] = None,
     bandwidth: Optional[float] = None,
