### 2022-08-08 48d543b
improve coverage of iclabel/label_components.py
