+++ mne_icalabel/iclabel/label_components.py
@@ -59,7 +59,7 @@ def iclabel_label_components(inst: Union[BaseRaw, BaseEpochs], ica: ICA, inplace
         for idx, (_, mne_label) in enumerate(ICLABEL_LABELS_TO_MNE.items()):
             auto_labels = np.argwhere(argmax_labels == idx)
             if mne_label not in ica.labels_:
-                ica.labels_[mne_label] = auto_labels
+                ica.labels_[mne_label] = list(auto_labels.flatten())
                 continue
             for comp in auto_labels:
                 if comp not in ica.labels_[mne_label]:
