### 2022-08-08 da50911
fix reshape_fortran as the tensor shape is always > 0
