### 2022-08-08 a012987
remove Optional[] for update_path as it does not prompt the user
