+++ mne_icalabel/datasets/icalabel/icalabel.py
@@ -1,4 +1,3 @@
-import os
 import shutil
 from functools import partial
 from typing import Optional
@@ -64,13 +63,12 @@ def data_path(
         download=download,
         processor=pooch.Unzip(extract_dir=f"./{folder_name}"),
     )
-    dpath = str(dpath)
 
-    # Do some wrangling to deal with nested directories
-    bad_name = os.path.join(dpath, "mne-testing-icalabel-data-main")
-    if os.path.isdir(bad_name):
-        os.rename(bad_name, dpath + ".true")
+    # do some wrangling to deal with nested directories
+    bad_name = dpath / "mne-testing-icalabel-data-main"
+    if bad_name.is_dir():
+        shutil.move(bad_name, dpath.with_suffix(".true"))
         shutil.rmtree(dpath)
-        os.rename(dpath + ".true", dpath)
+        shutil.move(dpath.with_suffix(".true"), dpath)
 
     return _mne_path(dpath)
