+++ pyproject.toml
@@ -174,6 +174,7 @@ source = [
 omit = [
     '**/__init__.py',
     'mne_icalabel/commands/*',
+    'mne_icalabel/_version.py',
     '**/tests/**',
 ]
 
