### 2022-08-08 8575d16
remove test for indent == sys.maxsize which would lead to memory error anyway
