### 2022-08-08 6ad7acc
add test for qt_version
