### 2022-08-08 89e7dae
improve coverage of label_components.py
