### 2022-08-08 ab6d3a3
better conversion to list
