+++ mne_icalabel/iclabel/label_components.py
@@ -57,9 +57,9 @@ def iclabel_label_components(inst: Union[BaseRaw, BaseEpochs], ica: ICA, inplace
 
         # add labels to the ICA instance
         for idx, (_, mne_label) in enumerate(ICLABEL_LABELS_TO_MNE.items()):
-            auto_labels = np.argwhere(argmax_labels == idx)
+            auto_labels = list(np.argwhere(argmax_labels == idx).flatten())
             if mne_label not in ica.labels_:
-                ica.labels_[mne_label] = list(auto_labels.flatten())
+                ica.labels_[mne_label] = auto_labels
                 continue
             for comp in auto_labels:
                 if comp not in ica.labels_[mne_label]:
