### 2022-08-08 83bf607
Revert "try different report config"
This reverts commit 3ffaf035049c1e33127b24cf0d642c2381439c71.
