+++ pyproject.toml
@@ -184,9 +184,3 @@ exclude_lines = [
     'if __name__ == .__main__.:',
 ]
 precision = 2
-omit = [
-    '**/__init__.py',
-    'mne_icalabel/commands/*',
-    'mne_icalabel/_version.py',
-    '**/tests/**',
-]
