### 2022-08-08 33856df
[ENH] get_topomap_array is now a private method  (#95)
* remove get_topomap_array from doc

* order alphabetically

* remove get_topomap_array from what's new

* make get_topomap_array private
