### 2022-08-08 a79b181
revert changes to topomap now covered in #95
