+++ mne_icalabel/features/tests/test_topomap.py
@@ -4,8 +4,7 @@ from mne.datasets import testing
 from mne.io import read_raw
 from mne.preprocessing import ICA
 
-from mne_icalabel.features import get_topomaps
-from mne_icalabel.features.topomap import _get_topomap_array
+from mne_icalabel.features import get_topomap_array, get_topomaps
 
 directory = testing.data_path() / "MEG" / "sample"
 raw = read_raw(directory / "sample_audvis_trunc_raw.fif", preload=False)
@@ -23,7 +22,7 @@ def test_topomap_defaults():
 
     # test single topo with fake data
     data = np.random.randint(1, 10, len(raw.ch_names))
-    topomap = _get_topomap_array(data, raw.info)
+    topomap = get_topomap_array(data, raw.info)
     assert isinstance(topomap, np.ndarray)
     assert topomap.shape == (64, 64)
 
@@ -44,7 +43,7 @@ def test_topomap_arguments(picks, res):
     assert topomaps.shape == (n_components, res, res)
 
     data = np.random.randint(1, 10, len(raw.ch_names))
-    topomap = _get_topomap_array(data, raw.info, res=res)
+    topomap = get_topomap_array(data, raw.info, res=res)
     assert isinstance(topomap, np.ndarray)
     assert topomap.shape == (res, res)
 
