+++ mne_icalabel/features/__init__.py
@@ -1,6 +1,6 @@
 """Features for the ICLabel"""
 
 from .psd import get_psds
-from .topomap import get_topomaps
+from .topomap import get_topomap_array, get_topomaps  # noqa: F401
 
 __all__ = ("get_psds", "get_topomaps")
