+++ mne_icalabel/features/topomap.py
@@ -49,7 +49,7 @@ def get_topomaps(
     topomaps = np.zeros((len(picks), res, res))
     for i, component in enumerate(picks):
         topo = np.flipud(
-            _get_topomap_array(
+            get_topomap_array(
                 data[component, :], ica.info, res, outlines, image_interp, border, extrapolate
             )
         )
@@ -61,7 +61,7 @@ def get_topomaps(
 
 
 @fill_doc
-def _get_topomap_array(
+def get_topomap_array(
     data: NDArray[float],
     pos: Info,
     res: int = 64,
