### 2022-08-08 bc5940c
add validation of an ICA instance
