+++ mne_icalabel/utils/tests/test_checks.py
@@ -5,7 +5,7 @@ from mne.io import RawArray
 from mne.preprocessing import ICA
 from mne.utils import requires_version
 
-from mne_icalabel.utils._checks import _check_qt_version, _validate_inst_and_ica
+from mne_icalabel.utils._checks import _check_qt_version, _validate_inst_and_ica, _validate_ica
 
 
 def test_validate_inst_and_ica():
@@ -29,6 +29,26 @@ def test_validate_inst_and_ica():
     ica.fit(raw)
     # test valid
     _validate_inst_and_ica(raw, ica)
+
+
+def test_validate_ica():
+    """Test that _validate_ica correctly raises."""
+    # check types
+    with pytest.raises(TypeError, match="ica must be an instance of ICA"):
+        _validate_ica(101)
+
+    # test unfitted
+    ica = ICA(n_components=3, method="picard")
+    with pytest.raises(RuntimeError, match="The provided ICA instance was not fitted."):
+        _validate_ica(ica)
+
+    # to avoid RuntimeWarning with fitting an unfiltered raw, let's fake the filter
+    raw = RawArray(np.random.randint(1, 10, (3, 1000)), create_info(3, 100, "eeg"))
+    with raw.info._unlock():
+        raw.info["highpass"] = 1.0
+    ica.fit(raw)
+    # test valid
+    _validate_ica(ica)
 
 
 # TODO: When tests of the GUI are improved and tests on all 4 Qt bindings, the
