+++ mne_icalabel/utils/tests/test_checks.py
@@ -5,7 +5,7 @@ from mne.io import RawArray
 from mne.preprocessing import ICA
 from mne.utils import requires_version
 
-from mne_icalabel.utils._checks import _check_qt_version, _validate_inst_and_ica, _validate_ica
+from mne_icalabel.utils._checks import _check_qt_version, _validate_ica, _validate_inst_and_ica
 
 
 def test_validate_inst_and_ica():
