### 2022-08-09 bebcea4
topomap: remove outlines as it does not affect the output
