+++ mne_icalabel/features/topomap.py
@@ -1,4 +1,4 @@
-from typing import Optional, Union
+from typing import Union
 
 import numpy as np
 from mne.channels.layout import _find_topomap_coords
@@ -6,6 +6,7 @@ from mne.defaults import _BORDER_DEFAULT, _EXTRAPOLATE_DEFAULT, _INTERPOLATION_D
 from mne.io import Info
 from mne.io.pick import _get_channel_types, _pick_data_channels, _picks_to_idx, pick_info
 from mne.preprocessing import ICA
+from mne.utils import _check_option, _validate_type
 from mne.viz.topomap import _check_extrapolate, _make_head_outlines, _setup_interp
 from numpy.typing import NDArray
 
@@ -18,7 +19,6 @@ def get_topomaps(
     ica: ICA,
     picks=None,
     res: int = 64,
-    outlines: Optional[str] = "head",
     image_interp: str = _INTERPOLATION_DEFAULT,  # 'cubic'
     border: Union[float, str] = _BORDER_DEFAULT,  # 'mean'
     extrapolate: str = _EXTRAPOLATE_DEFAULT,  # 'auto' -> 'head'
@@ -31,7 +31,6 @@ def get_topomaps(
         MNE `~mne.preprocessing.ICA` decomposition.
     %(picks_ica)s ``None`` (default) will pick all independent components in the order fitted.
     %(res_topomap)s
-    %(outlines_topomap)s
     %(image_interp_topomap)s
     %(border_topomap)s
     %(extrapolate_topomap)s
@@ -43,23 +42,26 @@ def get_topomaps(
     """
     _validate_ica(ica)
     picks = _picks_to_idx(ica.n_components_, picks)
+    _validate_type(res, "int", "res", "int")
+    if res <= 0:
+        raise ValueError(f"Argument 'res' should be a strictly positive integer. Provided '{res}' is invalid.")
+
     data = np.dot(
         ica.mixing_matrix_[:, : ica.n_components_].T,
         ica.pca_components_[: ica.n_components_],
     )
-    # Create an empty array of size (len(picks), res, res) for the topomap
+    # create an empty array of size (len(picks), res, res) for the topomap
     topomaps = np.zeros((len(picks), res, res))
     for i, component in enumerate(picks):
         topo = np.flipud(
             _get_topomap_array(
-                data[component, :], ica.info, res, outlines, image_interp, border, extrapolate
+                data[component, :], ica.info, res, image_interp, border, extrapolate
             )
         )
-        # Set NaN values to 0
         np.nan_to_num(topo, nan=0.0, copy=False)
-        # Standardize the values
+        # standardize
         topomaps[i, :, :] = topo / np.max(np.abs(topo))
-    return topomaps  # topographic map array for all the picked components (len(picks), res, res)
+    return topomaps  # topographic map array for all the picked components (len(picks), n_pixels, n_pixels)
 
 
 @fill_doc
@@ -67,7 +69,6 @@ def _get_topomap_array(
     data: NDArray[float],
     pos: Info,
     res: int = 64,
-    outlines: Optional[str] = "head",
     image_interp: str = _INTERPOLATION_DEFAULT,  # 'cubic'
     border: Union[float, str] = _BORDER_DEFAULT,  # 'mean'
     extrapolate: str = _EXTRAPOLATE_DEFAULT,  # 'head'
@@ -81,7 +82,6 @@ def _get_topomap_array(
     pos : Info
         Instance of `mne.Info` with the montage associated with the ``(n_channels,)`` points.
     %(res_topomap)s
-    %(outlines_topomap)s
     %(image_interp_topomap)s
     %(border_topomap)s
[…]
