+++ mne_icalabel/utils/_docs.py
@@ -37,13 +37,6 @@ image_interp : str
     The image interpolation to be used. All matplotlib options are
     accepted."""
 docdict[
-    "outlines_topomap"
-] = """
-outlines : 'head' | 'skirt' | None
-    The outlines of the head. If ``'head'``, the default MNE head scheme. If
-    ``'skirt'``, the default MNE head scheme but sensors are allowed to be positioned
-    outside of the head circle. If None, the outline is disabled."""
-docdict[
     "res_topomap"
 ] = """
 res : int
