### 2022-08-09 8956959
topomap: improvements to the structure and better comments
