+++ mne_icalabel/features/topomap.py
@@ -6,7 +6,7 @@ from mne.defaults import _BORDER_DEFAULT, _EXTRAPOLATE_DEFAULT, _INTERPOLATION_D
 from mne.io import Info
 from mne.io.pick import _get_channel_types, _pick_data_channels, _picks_to_idx, pick_info
 from mne.preprocessing import ICA
-from mne.utils import _check_option, _validate_type
+from mne.utils import _validate_type
 from mne.viz.topomap import _check_extrapolate, _make_head_outlines, _setup_interp
 from numpy.typing import NDArray
 
@@ -44,23 +44,22 @@ def get_topomaps(
     picks = _picks_to_idx(ica.n_components_, picks)
     _validate_type(res, "int", "res", "int")
     if res <= 0:
-        raise ValueError(f"Argument 'res' should be a strictly positive integer. Provided '{res}' is invalid.")
+        raise ValueError(
+            f"Argument 'res' should be a strictly positive integer. Provided '{res}' is invalid."
+        )
+    # image_interp, border are validated by _setup_interp
+    # extrapolate is validated by _check_extrapolate
 
     data = np.dot(
         ica.mixing_matrix_[:, : ica.n_components_].T,
         ica.pca_components_[: ica.n_components_],
     )
-    # create an empty array of size (len(picks), res, res) for the topomap
+    # create an empty array of size (len(picks), n_pixels, n_pixels) for the topomap
     topomaps = np.zeros((len(picks), res, res))
     for i, component in enumerate(picks):
-        topo = np.flipud(
-            _get_topomap_array(
-                data[component, :], ica.info, res, image_interp, border, extrapolate
-            )
+        topomaps[i, :, :] = _get_topomap_array(
+            data[component, :], ica.info, res, image_interp, border, extrapolate
         )
-        np.nan_to_num(topo, nan=0.0, copy=False)
-        # standardize
-        topomaps[i, :, :] = topo / np.max(np.abs(topo))
     return topomaps  # topographic map array for all the picked components (len(picks), n_pixels, n_pixels)
 
 
@@ -115,4 +114,7 @@ def _get_topomap_array(
     outlines = _make_head_outlines(sphere, pos, None, (0.0, 0.0))
     extent, Xi, Yi, interp = _setup_interp(pos, res, image_interp, extrapolate, outlines, border)
     interp.set_values(data)
-    return interp.set_locations(Xi, Yi)()  # Zi, topomap of shape (n_pixels, n_pixels)
+    topomap = np.flipud(interp.set_locations(Xi, Yi)())  # Zi, shape (n_pixels, n_pixels)
+    np.nan_to_num(topomap, nan=0.0, copy=False)
+    topomap = topomap / np.max(np.abs(topomap))  # standardize
+    return topomap  # (n_pixels, n_pixels)
