### 2022-08-09 0e517fc
retrieve data only for the picked ICs
