### 2022-08-09 1bd4a05
topomap: support ICA with multiple channel types
