+++ mne_icalabel/features/tests/test_topomap.py
@@ -2,30 +2,32 @@ import numpy as np
 import pytest
 from mne.datasets import testing
 from mne.io import read_raw
+from mne.io.pick import _get_channel_types, _pick_data_channels, _picks_to_idx, pick_info
 from mne.preprocessing import ICA
 
 from mne_icalabel.features import get_topomaps
-from mne_icalabel.features.topomap import _get_topomap_array
 
 directory = testing.data_path() / "MEG" / "sample"
 raw = read_raw(directory / "sample_audvis_trunc_raw.fif", preload=False)
-raw.pick_types(eeg=True)
+raw.pick_types(eeg=True, meg=True)
 raw.load_data()
 ica = ICA(n_components=5, method="picard")
 ica.fit(raw)
+ica_eeg = ICA(n_components=5, method="picard")
+ica_eeg.fit(raw.copy().pick_types(eeg=True))
 
 
-def test_topomap_defaults():
+@pytest.mark.parametrize("ica", (ica, ica_eeg))
+def test_topomap_defaults(ica):
     """Test scalp topography array generation"""
     topomaps = get_topomaps(ica, picks=None)
-    assert isinstance(topomaps, np.ndarray)
-    assert topomaps.shape == (ica.n_components_, 64, 64)
-
-    # test single topo with fake data
-    data = np.random.randint(1, 10, len(raw.ch_names))
-    topomap = _get_topomap_array(data, raw.info)
-    assert isinstance(topomap, np.ndarray)
-    assert topomap.shape == (64, 64)
+    assert isinstance(topomaps, dict)
+    for topomaps_ in topomaps.values():
+        assert isinstance(topomaps_, np.ndarray)
+        assert topomaps_.shape == (ica.n_components_, 64, 64)
+    ch_picks = _pick_data_channels(ica.info, exclude=())
+    ch_types = _get_channel_types(pick_info(ica.info, ch_picks), unique=True)
+    assert sorted(topomaps) == sorted(ch_types)
 
 
 @pytest.mark.parametrize(
@@ -34,19 +36,53 @@ def test_topomap_defaults():
 def test_topomap_arguments(picks, res):
     """Test arguments that influence the output shape."""
     topomaps = get_topomaps(ica, picks=picks, res=res)
-    assert isinstance(topomaps, np.ndarray)
-    if isinstance(picks, int):
-        n_components = 1
-    elif isinstance(picks, slice):
-        n_components = len(range(*picks.indices(ica.n_components_)))
-    else:
-        n_components = len(picks)
-    assert topomaps.shape == (n_components, res, res)
+    assert isinstance(topomaps, dict)
+    ic_picks = _picks_to_idx(ica.n_components_, picks)
+    for topomaps_ in topomaps.values():
+        assert isinstance(topomaps_, np.ndarray)
+        assert topomaps_.shape == (ic_picks.size, res, res)
 
-    data = np.random.randint(1, 10, len(raw.ch_names))
-    topomap = _get_topomap_array(data, raw.info, res=res)
-    assert isinstance(topomap, np.ndarray)
-    assert topomap.shape == (res, res)
+
+def test_interpolation_arguments():
+    """Test arguments that influence the interpolation."""
+    # default
+    topomaps1 = get_topomaps(ica)
+
+    # image_interp
+    topomaps2 = get_topomaps(ica, image_interp="linear")
+    topomaps3 = get_topomaps(ica, image_interp="nearest")
+    for topomaps in (topomaps2, topomaps3):
+        assert sorted(topomaps1) == sorted(topomaps)
+        for ch_type in topomaps1:
[…]
