### 2022-08-09 edfce63
topomap: better tests
