### 2022-08-09 e9736d5
remove unnecessary copy
