### 2022-08-09 6e99557
remove appendix to picks_ica which should be added to the MNE docdict
