### 2022-08-09 09d3673
topomap: move documentation for picks to the docstring to remove the 'visualize' mention
