### 2022-08-09 cf2323a
simplify numpydoc x-ref
