### 2022-08-09 dabac44
replace np.ndarray with array in doc
