### 2022-08-09 48bfeb2
better numpydoc_xref_aliases because ="all" is supported but issues a warning
