### 2022-08-09 4d0c292
simplify numpydoc config
