### 2022-08-09 cef4630
[MNT] Improve coverage (#94)
* improve coverage of iclabel/label_components.py

* fix reshape_fortran as the tensor shape is always > 0

* add type-hints

* remove Optional[] for update_path as it does not prompt the user

* simplify

* Revert "simplify"

This reverts commit db6c7cac2be9cd7295b90891fc07b0b5e195bb2f.

* simpler

* improve datasets coverage

* fix style

* ignore _version.py file and add TODO item in _version.py

* remove test for indent == sys.maxsize which would lead to memory error anyway

* improve coverage of utils/_docs.py

* better comment

* add test for qt_version

* fi
[…]
