+++ doc/conf.py
@@ -1,4 +1,5 @@
 """Configure details for documentation with sphinx."""
+
 import os
 import subprocess
 import sys
@@ -17,32 +18,31 @@ import mne_icalabel  # noqa: E402
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 curdir = os.path.dirname(__file__)
-sys.path.append(os.path.abspath(os.path.join(curdir, '..')))
-sys.path.append(os.path.abspath(os.path.join(curdir, '..', 'mne_icalabel')))
-sys.path.append(os.path.abspath(os.path.join(curdir, 'sphinxext')))
+sys.path.append(os.path.abspath(os.path.join(curdir, "..")))
+sys.path.append(os.path.abspath(os.path.join(curdir, "..", "mne_icalabel")))
+sys.path.append(os.path.abspath(os.path.join(curdir, "sphinxext")))
 
 # -- General configuration ------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
 #
-needs_sphinx = '4.0'
+needs_sphinx = "4.0"
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.autosummary',
-    'sphinx.ext.doctest',
-    'sphinx.ext.intersphinx',
-    'sphinx_autodoc_typehints',
-    'sphinx.ext.mathjax',
-    'sphinx.ext.viewcode',
-    'sphinx_gallery.gen_gallery',
-    'sphinxcontrib.bibtex',
-    'numpydoc',
-    'sphinx_copybutton',
-    'gh_substitutions',  # custom extension, see sphinxext/gh_substitutions.py
+    "sphinx.ext.autodoc",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.doctest",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.mathjax",
+    "sphinx.ext.viewcode",
+    "sphinx_gallery.gen_gallery",
+    "sphinxcontrib.bibtex",
+    "numpydoc",
+    "sphinx_copybutton",
+    "gh_substitutions",  # custom extension, see sphinxext/gh_substitutions.py
 ]
 
 # configure sphinx-copybutton
@@ -53,25 +53,19 @@ copybutton_prompt_is_regexp = True
 # -- sphinx.ext.autosummary
 autosummary_generate = True
 
-autodoc_default_options = {'inherited-members': None}
-autodoc_typehints = 'signature'
+autodoc_default_options = {"inherited-members": None}
+autodoc_typehints = "none"
 
 # prevent jupyter notebooks from being run even if empty cell
 # nbsphinx_execute = 'never'
 # nbsphinx_allow_errors = True
 
-error_ignores = {
-    # These we do not live by:
-    'GL01',  # Docstring should start in the line immediately after the quotes
-    'EX01', 'EX02',  # examples failed (we test them separately)
-    'ES01',  # no extended summary
-    'SA01',  # no see also
-    'YD01',  # no yields section
-    'SA04',  # no description in See Also
-    'PR04',  # Parameter "shape (n_channels" has no type
-    'RT02',  # The first line of the Returns section should contain only the type, unless multiple values are being returned  # noqa
-    # XXX should also verify that | is used rather than , to separate params
-    # XXX should maybe also restore the parameter-desc-length < 800 char check
[…]
