+++ mne_icalabel/label_components.py
@@ -30,7 +30,7 @@ def label_components(inst: Union[BaseRaw, BaseEpochs], ica: ICA, method: str):
     component_dict : dict
         A dictionary with the following fields:
 
-        - 'y_pred_proba' : np.ndarray of shape (n_components,)
+        - 'y_pred_proba' : array of shape (n_components,)
             Estimated predicted probability of the output class
             for each independent component.
         - 'labels': list of shape (n_components,)
@@ -49,9 +49,7 @@ def label_components(inst: Union[BaseRaw, BaseEpochs], ica: ICA, method: str):
     - 'other'
     """
     _validate_type(method, str, "method")
-    _check_option("method", method, ICALABEL_METHODS)
-    if ICALABEL_METHODS[method] is None:
-        raise RuntimeError('Cannot pass in "manual" as a method.')
+    _check_option("method", method, [elt for elt in ICALABEL_METHODS if elt != "manual"])
 
     _validate_inst_and_ica(inst, ica)
     labels_pred_proba = ICALABEL_METHODS[method](inst, ica)  # type: ignore
