+++ doc/whats_new_previous_releases.rst
@@ -16,15 +16,15 @@ added an experimental feature for helping add ICA component annotations to BIDS-
 Enhancements
 ~~~~~~~~~~~~
 
-- Add functions for annotating and labeling ICA components in BIDS format :func:`mne_icalabel.annotation.write_components_tsv`, :func:`mne_icalabel.annotation.mark_component` by `Adam Li`_ (:gh:`60`)
+- Add functions for annotating and labeling ICA components in BIDS format :func:`mne_icalabel.annotation.write_components_tsv`, :func:`mne_icalabel.annotation.mark_component` by `Adam Li`_ (:pr:`60`)
 
 Bug
 ~~~
 
-- Fix shape of ``'y_pred_proba'`` output from :func:`mne_icalabel.label_components` by `Mathieu Scheltienne`_ (:gh:`36`)
-- Add a warning if the ICA decomposition provided does not match the expected decomposition by ``ICLabel``  by `Mathieu Scheltienne`_ (:gh:`42`)
-- Fix extraction of PSD feature from ``ICLabel`` model on epochs by `Mathieu Scheltienne`_ (:gh:`64`)
-- Fix ICLabel topographic features on ICA fitted with a channel selection performed by ``picks`` by `Mathieu Scheltienne`_ (:gh:`68`)
+- Fix shape of ``'y_pred_proba'`` output from :func:`mne_icalabel.label_components` by `Mathieu Scheltienne`_ (:pr:`36`)
+- Add a warning if the ICA decomposition provided does not match the expected decomposition by ``ICLabel``  by `Mathieu Scheltienne`_ (:pr:`42`)
+- Fix extraction of PSD feature from ``ICLabel`` model on epochs by `Mathieu Scheltienne`_ (:pr:`64`)
+- Fix ICLabel topographic features on ICA fitted with a channel selection performed by ``picks`` by `Mathieu Scheltienne`_ (:pr:`68`)
 
 API
 ~~~
@@ -55,7 +55,7 @@ People who contributed to this release (in alphabetical order):
 
 * `Adam Li`_
 * `Jacob Feitelberg`_
-* `Anand Saini`_ 
+* `Anand Saini`_
 * `Mathieu Scheltienne`_
 
 .. include:: authors.inc
