### 2022-08-09 b26cdec
[DOC] Remove typing-extensions and replace outdated gh-substitutions by sphinx-issues (#96)
* simpler

* configure sphinx-issues
