### 2022-08-10 c442c23
[DOCS] Add docs for gui (#97)
* Add docs for gui

* Fix docs
