+++ mne_icalabel/gui/__init__.py
@@ -6,11 +6,11 @@ def label_ica_components(inst, ica: ICA, show: bool = True, block: bool = False)
 
     Parameters
     ----------
-    inst : : Raw | Epochs
+    inst : Raw | Epochs
         `~mne.io.Raw` or `~mne.Epochs` instance used to fit the `~mne.preprocessing.ICA`
         decomposition.
     ica : ICA
-        The ICA object fitted on `inst`.
+        The ICA object fitted on ``inst``.
     show : bool
         Show the GUI if True.
     block : bool
