+++ doc/conf.py
@@ -175,8 +175,9 @@ html_sidebars = {
 
 html_context = {
     "versions_dropdown": {
-        "dev": "v0.3 (devel)",
-        "stable": "v0.2",
+        "dev": "v0.4 (devel)",
+        "stable": "v0.3",
+        "v0.2": "v0.2",
         "v0.1": "v0.1",
     },
 }
