+++ .circleci/config.yml
@@ -243,6 +243,7 @@ workflows:
             branches:
               only:
                 - main
+                - maint/0.3
                 - maint/0.2
 
   main:
