+++ .circleci/config.yml
@@ -218,6 +218,11 @@ jobs:
               fingerprints:
                 - "b5:1e:a1:6d:8d:48:f2:8f:dd:bd:2d:66:a9:30:fe:b9"
           - run:
+              name: Check workspace
+              command: |
+                ls -al ./;
+                ls -al doc/_build/;
+          - run:
               # push built doc into the `dev` directory on the `gh-pages` branch
               name: Deploy doc to gh-pages branch
               command: |
