+++ .circleci/config.yml
@@ -105,16 +105,14 @@ jobs:
                python -c "import mne; level = mne.get_config('MNE_LOGGING_LEVEL'); assert level.lower() == 'info', repr(level)"
                python -c "import numpy; numpy.show_config()"
         - run:
-            name: Verify build type
-            command: |
-              echo "PATTERN=$(cat pattern.txt)"
-              echo "BUILD=$(cat build.txt)"
-        - run:
             name: Build documentation
             command: |
-              # cd doc
-              # make html
-              PATTERN=$(cat pattern.txt) make -C doc $(cat build.txt);
+              cd doc
+              if [ "${CIRCLE_BRANCH}" == "main" ]; then
+                make html
+              else
+                make html_stable
+              fi;
         # Save the example test results
         - store_test_results:
             path: doc/_build/test-results
