### 2022-08-11 1450288
[RELEASE] V0.3 (#98)
* Adding stable docs

* Adding commit to release v0.3

* Added

* Adding update to try to trigger CI

* Try again

* Try again

* Fix

* Try again

* Fix
