+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = 'setuptools.build_meta'
 
 [project]
 name = 'mne-icalabel'
-version = '0.4dev0'
+version = '0.3.1'
 description = 'MNE-ICALabel: Automatic labeling of ICA components from MEG, EEG and iEEG data with MNE.'
 readme = 'README.md'
 license = {file = 'LICENSE'}
