+++ pyproject.toml
@@ -121,7 +121,7 @@ include = ['mne_icalabel*']
 exclude = ['mne_icalabel*tests']
 
 [tool.setuptools.package-data]
-"mne_icalabel.iclabel" = ["asserts/*.pt"]
+"mne_icalabel.iclabel" = ["assets/*.pt"]
 
 [tool.black]
 line-length = 100
