+++ .github/workflows/unit_tests.yml
@@ -158,7 +158,7 @@ jobs:
       - name: Run pytest  # headless via Xvfb on linux
         uses: GabrielBB/xvfb-action@v1.6
         with:
-          run: pytest main/mne_icalabel --cov=mne_icalabel --cov-report=xml --cov-config=main/pyproject.toml
+          run: pytest main/mne_icalabel --cov=main/mne_icalabel --cov-report=xml --cov-config=main/pyproject.toml
       - name: Upload coverage stats to codecov
         if: ${{ matrix.os == 'ubuntu' && matrix.python-version == '3.10' && matrix.mne == 'stable' }}
         uses: codecov/codecov-action@v3
