+++ .github/workflows/unit_tests.yml
@@ -86,6 +86,7 @@ jobs:
       - name: Build package
         run: python -m build
       - name: Upload package distribution files
+        if: ${{ matrix.os == 'ubuntu' && matrix.python-version == '3.10' }}
         uses: actions/upload-artifact@v3
         with:
           name: package
