### 2022-08-15 #4484 @emdupre
Hi @adam2392 ,

Just to update you on this : I've reached out to @hvgazula several times but unfortunately have not received a response. Hopefully he's OK, but we'll now need to move forward without his review. I'm in the process of securing a second reviewer who could provide their input on an accelerated timeline so that we can move this forward.

I'm sorry for the delay in reviewing this work, and I appreciate your patience.

To try and save time where we can, I'll go ahead and make a few editorial comments here and on the software repo. Please let me know if you have any questions about these, or anything else at this point.
