### 2022-08-17 b91ac87
fix x-ref to bool and remove unused warnings import
