### 2022-08-17 969109f
Update mne_icalabel/annotation/bids.py
Co-authored-by: Mathieu Scheltienne <mathieu.scheltienne@gmail.com>
