### 2022-08-18 #4484 @emdupre
As an update, @adswa has agreed to review this submission on an accelerated timeline, so I'm adding her as a reviewer now. Thank you, @adswa !

Adapting relevant information from the top comment:

> Your review will be checklist based. You will have a separate checklist that you should update when carrying out your review.
> First, you need to run this command in a separate comment to create the checklist:

```@editorialbot generate my checklist```

> The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let @emdupre know.
