### 2022-08-18 mne-tools/mne-icalabel#102 @adswa
[DOC] License link in documentation 404's
Hello :wave: :)
I read through the documentation and notived that the "License" header in the [documentation](https://mne.tools/mne-icalabel/stable/index.html#) points to https://raw.githubusercontent.com/mne-tools/mne-icalabel/blob/main/LICENSE, which does not resolve. The correct link would be https://raw.githubusercontent.com/mne-tools/mne-icalabel/main/LICENSE.

Here's a patch:

```diff
diff --git a/doc/index.rst b/doc/index.rst
index 2a33c25..257ed39 100644
--- a/doc/index.rst
+++ b/doc/index.rst
@@ -32,7 +32,7 @@ Contents
    :hidden:
    :caption: Development

-   License <https://raw.githubusercontent.com/mne-tools/mne-icalabel/blob/main/LICENSE>
+   License <https://raw.githubusercontent.com/mne-tools/mne-icalabel/main/LICENSE>
    Contributing <https://github.com/mne-tools/mne-icalabel/blob/main/CONTRIBUTING.md>

 License
```

Found while doing https://github.com/openjournals/joss-reviews/issues/4484

PS: I noticed that your repositories "About" overview section on the right links to the development version instead of the stable version of the documentation - is this intentional? Not in any way a problem, adding here as an FYI in case its an unintentional pre-first-release residual, because the empty "Whats new" page in the development version caught my eye and I'm used to redirection to the stable docs by mne-python. Ignore if this choice intentional or preferred.
