### 2022-08-18 #4484 @adswa
Hey everybody, happy to be on board! :wave:
I got a head start into the review, and had time to look at almost all review criteria already. Overall, congrats on a well-made piece of software and this cool addition into the mne ecosystem! I tested all installation methods and went through the examples in the documentation on a Debian system. I'm a big fan of the examples in MNEs documentation and was happy to see this software create similar walkthroughs. I got to run the test suite locally on a Debian system, too, and the automated CIs look good. I found contributing and license information easily in the source repo and in the docs (filed only https://github.com/mne-tools/mne-icalabel/issues/102 for a small bug). I plan to give the tool a test run on two different systems tomorrow, and might leave a comment or two about the paper structure, but you should have my complete review by tomorrow evening.
