### 2022-08-19 mne-tools/mne-icalabel#104 @adswa
[GUI] Missing or non-expandable ICA properties
#### Describe the bug
The GUI seems to lack some ICA property plots shown in the tutorial, or they exists, but can't be expanded.

#### Steps to reproduce
I followed the [tutorial](https://mne.tools/mne-icalabel/dev/auto_examples/label_components.html#sphx-glr-auto-examples-label-components-py), using a Python 3.9 environment on a Windows 11 and a Debian bookworm system.

#### Expected results
I expected to see the GUI include all plots shown in the figure in the tutorial, including the epochs image, ERP/ERF, power spectrum, and epoch variance (as shown in images
``doc/auto_examples/images/sphx_glr_label_components_00{1..4}.png`` in the examples).

#### Actual results
On both machines, this is how the GUI looks by default, lacking the images from  ``doc/auto_examples/images/sphx_glr_label_components_003.png`` (no external screen connected, display resolution of 2560x1440 (16:9) (Debian) and 1920x1080 (Windows)):

Debian:
![Screenshot from 2022-08-19 09-03-07](https://user-images.githubusercontent.com/29738718/185564983-425f854d-4868-431c-8977-79122c772b75.png)

Windows:
(pasting from a different machine)

This is how the GUI looks in full screen:
Debian:
![Screenshot from 2022-08-19 09-07-24](https://user-images.githubusercontent.com/29738718/185565047-a53725c9-2bd7-4bea-ace7-5a6622a834ac.png)

Windows:
(pasting from a different machine)

The plots might be present  underneath the topography, but the GUI components can't be changed in size individually so I can't reveal them if they are there.
