### 2022-08-19 mne-tools/mne-icalabel#103 @mscheltienne
Yes, we took the same approach as `mne-qt-browser`: install `qtpy` but don't install one of the 4 possibles `qt bindings`. c.f. their [requirement file](https://github.com/mne-tools/mne-qt-browser/blob/main/requirements.txt). It is up to the user to install in the environment one of `PyQt5`, `PyQt6`, `PySide2`, or `Pyside6`.

The difference with `mne-qt-browser` is that we raise if the bindings are missing, while the browser will swap the backend to `matplotlib`.

The error message should be improved, it's not super explicit, and maybe also let's add a note in the tutorial. Would that be acceptable to resolve this issue?
