### 2022-08-19 mne-tools/mne-icalabel#104 @mscheltienne
They are not present. We only included those 3 features for now. Also, you might have noticed that for now, each time you select a new IC, the previous `mne-qt-browser` (bottom right) is closed and a new one is opened and placed in the window. Not ideal.

This GUI is a very recent addition, it's an early version. The layout and the size of the widgets are not great. The features included will be expanded. It's a work in progress for now.
