### 2022-08-19 mne-tools/mne-icalabel#102 @mscheltienne
Thanks for the review, that should be fixed, I'll check when it's deployed. For the link in the "About" overview section, @adam2392 do you have any preference?
