+++ doc/index.rst
@@ -32,7 +32,7 @@ Contents
    :hidden:
    :caption: Development
 
-   License <https://raw.githubusercontent.com/mne-tools/mne-icalabel/blob/main/LICENSE>
+   License <https://raw.githubusercontent.com/mne-tools/mne-icalabel/main/LICENSE>
    Contributing <https://github.com/mne-tools/mne-icalabel/blob/main/CONTRIBUTING.md>
 
 License
