### 2022-08-19 #4484 @adswa
Alright, I'm done with my review. I filed a range of general observations I made when trying it out a bit more (https://github.com/mne-tools/mne-icalabel/issues/103, https://github.com/mne-tools/mne-icalabel/issues/104, https://github.com/mne-tools/mne-icalabel/issues/107, https://github.com/mne-tools/mne-icalabel/issues/102), of which I deem none required to close in order to accept this submission. The only thing I would ask the authors for is to resolve my confusion stated in https://github.com/mne-tools/mne-icalabel/issues/105 about the data requirements, and if my assumptions are not wrong, add some clarifying remarks to documentation, warnings, or paper as they see appropriate to help users select and adopt the tool more efficiently. Other than that, I'm very happy with the submission, and ready to recommend "accept".
