+++ examples/iclabel_automatic_artifact_correction_ica.py
@@ -6,7 +6,7 @@ Repairing artifacts with ICA automatically using ICLabel Model
 ==============================================================
 
 This tutorial covers automatically repairing signals using ICA with
-the ICLabel model :footcite:`iclabel2019`, which originates in EEGLab.
+the ICLabel model\ :footcite:`iclabel2019`, which originates in EEGLab.
 For conceptual background on ICA, see :ref:`this scikit-learn tutorial
 <sphx_glr_auto_examples_decomposition_plot_ica_blind_source_separation.py>`.
 For a basic understanding of how to use ICA to remove artifacts, see `the
@@ -14,33 +14,34 @@ tutorial
 <https://mne.tools/stable/auto_tutorials/preprocessing/40_artifact_correction_ica.html>`_
 in MNE-Python.
 
-We begin as always by importing the necessary Python modules and loading some
-:ref:`example data <sample-dataset>`. Because ICA can be computationally
-intense, we'll also crop the data to 60 seconds; and to save ourselves from
-repeatedly typing ``mne.preprocessing`` we'll directly import a few functions
-and classes from that submodule.
+ICLabel was designed to label ICs fitted with an extended infomax algorithm on
+EEG data referenced to a common average. This model was not validated on other
+type of dataset.
 
-Note: this example involves running the ICA algorithm, which requires `scikit-learn`_
-to be installed. Please install this optional dependency before running the example.
+.. note::
+    This example involves running the ICA Infomax algorithm, which requires
+    `scikit-learn`_ to be installed. Please install this optional dependency before
+    running the example.
 """
 
 # %%
-
-import os
+# We begin as always by importing the necessary Python modules and loading some
+# :ref:`example data <sample-dataset>`. Because ICA can be computationally
+# intense, we'll also crop the data to 60 seconds; and to save ourselves from
+# repeatedly typing ``mne.preprocessing`` we'll directly import a few functions
+# and classes from that submodule.
 
 import mne
 from mne.preprocessing import ICA
 
 from mne_icalabel import label_components
 
-sample_data_folder = mne.datasets.sample.data_path()
-sample_data_raw_file = os.path.join(
-    sample_data_folder, "MEG", "sample", "sample_audvis_filt-0-40_raw.fif"
-)
+sample_data_folder = mne.datasets.sample.data_path() / "MEG" / "sample"
+sample_data_raw_file = sample_data_folder / "sample_audvis_raw.fif"
 raw = mne.io.read_raw_fif(sample_data_raw_file)
 
-# Here we'll crop to 60 seconds and drop gradiometer channels for speed
-raw.crop(tmax=60.0).pick_types(meg="mag", eeg=True, stim=True, eog=True)
+# we'll crop to 60 seconds and drop MEG channels
+raw.crop(tmax=60.0).pick_types(eeg=True, stim=True, eog=True)
 raw.load_data()
 
 # %%
@@ -62,13 +63,8 @@ raw.load_data()
 # Let's begin by visualizing the artifacts that we want to repair. In this
 # dataset they are big enough to see easily in the raw data:
 
-# Note: for this example, we are using ICLabel which has only
-# been validated and works for EEG systems with less than 32 electrodes.
-raw = raw.pick_types(eeg=True, eog=True, ecg=True, emg=True)
-
 # pick some channels that clearly show heartbeats and blinks
-regexp = r"(EEG 00.)"
-artifact_picks = mne.pick_channels_regexp(raw.ch_names, regexp=regexp)
+artifact_picks = mne.pick_channels_regexp(raw.ch_names, regexp=r"(EEG 00.)")
 raw.plot(order=artifact_picks, n_channels=len(artifact_picks), show_scrollbars=False)
 
 # %%
[…]
