### 2022-08-19 db67ae7
add note in the iclabel_label_components docstring and add the function to the documented functions
