### 2022-08-19 mne-tools/mne-icalabel#103 @adswa
> The error message should be improved, it's not super explicit, and maybe also let's add a note in the tutorial. Would that be acceptable to resolve this issue?

sure. thanks for the delineation. :)
