### 2022-08-19 a60ad09
better documentaiton of'label_components' entry-point
