### 2022-08-19 4c39e28
use same note everywhere and add information about filter
