### 2022-08-19 mne-tools/mne-icalabel#86 @TomDonoghue
This update looks good to me!
