### 2022-08-20 a7a4c94
clean-up configuration
