+++ doc/conf.py
@@ -20,11 +20,24 @@ curdir = os.path.dirname(__file__)
 sys.path.append(os.path.abspath(os.path.join(curdir, "..")))
 sys.path.append(os.path.abspath(os.path.join(curdir, "..", "mne_icalabel")))
 
-# -- General configuration ------------------------------------------------
+# -- project information -----------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+
+# General information about the project.
+project = "MNE-ICALabel"
+author = "Adam Li"
+td = date.today()
+copyright = f"2021-{td.year}, MNE Developers. Last updated on {td.isoformat()}"
+
+# -- general configuration ------------------------------------------------
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 # If your documentation needs a minimal Sphinx version, state it here.
-#
 needs_sphinx = "4.0"
+
+# The document name of the “root” document, that is, the document that contains
+# the root toctree directive.
+root_doc = "index"
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -43,44 +56,33 @@ extensions = [
     "sphinx_issues",
 ]
 
-# configure sphinx-copybutton
-copybutton_prompt_text = r">>> |\.\.\. |\$ "
-copybutton_prompt_is_regexp = True
+# Sphinx will warn about all references where the target cannot be found.
+nitpicky = True
+nitpick_ignore = []
 
-# generate autosummary even if no references
-# -- sphinx.ext.autosummary
+# The name of a reST role (builtin or Sphinx extension) to use as the default
+# role, that is, for text marked up `like this`. This can be set to 'py:obj' to
+# make `filter` a cross-reference to the Python function “filter”.
+default_role = "py:obj"
+
+# -- autosummary -------------------------------------------------------------
 autosummary_generate = True
 
-autodoc_default_options = {"inherited-members": None}
+# -- autodoc -----------------------------------------------------------------
+autoclass_content = "class"
 autodoc_typehints = "none"
+autodoc_member_order = "groupwise"
+autodoc_warningiserror = True
+autodoc_default_options = {"inherited-members": None}
 
-# prevent jupyter notebooks from being run even if empty cell
-# nbsphinx_execute = 'never'
-# nbsphinx_allow_errors = True
+# -- numpydoc ----------------------------------------------------------------
 
-error_ignores = {  # These we do not live by:
-    "GL01",  # docstring should start in the line immediately after the quotes
-    "EX01",  # section 'Examples' not found
-    "ES01",  # no extended summary found
-    "SA01",  # section 'See Also' not found
-    "RT02",  # The first line of the Returns section should contain only the type, unless multiple values are being returned  # noqa
-}
-
-# -- numpydoc
-# Below is needed to prevent errors
-numpydoc_xref_param_type = True
+# needed to prevent errors
 numpydoc_class_members_toctree = False
 numpydoc_attributes_as_param_list = True
-numpydoc_use_blockquotes = True
-numpydoc_xref_ignore = {
-    "of",
-    "shape",
-    "n_components",
-    "n_pixels",
-    "n_classes",
-    "instance",
-    "ICAComponentLabeler",
-}
+
+# x-ref
+numpydoc_xref_param_type = True
 numpydoc_xref_aliases = {
     # Python
     "Path": "pathlib.Path",
@@ -91,6 +93,26 @@ numpydoc_xref_aliases = {
     "Info": "mne.Info",
     "Raw": "mne.io.Raw",
 }
[…]
