### 2022-08-20 d61fabb
replace viewcode with linkcode (#109)
