+++ doc/conf.py
@@ -32,6 +32,17 @@ author = "Adam Li"
 td = date.today()
 copyright = f"2021-{td.year}, MNE Developers. Last updated on {td.isoformat()}"
 
+# The version info for the project you're documenting, acts as replacement for
+# |version| and |release|, also used in various other places throughout the
+# built documents.
+#
+# The short X.Y version.
+version = mne_icalabel.__version__
+# The full version, including alpha/beta/rc tags.
+release = version
+
+gh_url = "https://github.com/mne-tools/mne-icalabel"
+
 # -- general configuration ------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
@@ -59,6 +70,10 @@ extensions = [
     "sphinx_issues",
 ]
 
+source_suffix = ".rst"
+templates_path = ["_templates"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "**.ipynb_checkpoints"]
+
 # Sphinx will warn about all references where the target cannot be found.
 nitpicky = True
 nitpick_ignore = []
@@ -67,6 +82,56 @@ nitpick_ignore = []
 # role, that is, for text marked up `like this`. This can be set to 'py:obj' to
 # make `filter` a cross-reference to the Python function “filter”.
 default_role = "py:obj"
+
+# -- options for HTML output -------------------------------------------------
+
+# HTML options (e.g., theme)
+# see: https://sphinx-bootstrap-theme.readthedocs.io/en/latest/README.html
+# Clean up sidebar: Do not show "Source" link
+html_show_sourcelink = False
+html_copy_source = False
+
+html_theme = "pydata_sphinx_theme"
+
+# Add any paths that contain templates here, relative to this directory.
+html_static_path = ["_static"]
+html_css_files = ["style.css"]
+
+# Theme options are theme-specific and customize the look and feel of a theme
+# further.  For a list of options available for each theme, see the
+# documentation.
+html_theme_options = {
+    "icon_links": [
+        dict(
+            name="GitHub",
+            url=gh_url,
+            icon="fab fa-github-square",
+        ),
+    ],
+    "use_edit_page_button": False,
+    "navigation_with_keys": False,
+    "show_toc_level": 1,
+    "navbar_end": ["theme-switcher", "version-switcher", "navbar-icon-links"],
+}
+# Custom sidebar templates, maps document names to template names.
+html_sidebars = {
+    "index": ["search-field.html"],
+}
+
+html_context = {
+    "versions_dropdown": {
+        "dev": "v0.4 (devel)",
+        "stable": "v0.3",
+        "v0.2": "v0.2",
+        "v0.1": "v0.1",
+    },
+}
+
+# variables to pass to HTML templating engine
+html_context = {
+    'pygment_light_style': 'tango',
+    'pygment_dark_style': 'native',
+}
 
 # -- autosummary -------------------------------------------------------------
 autosummary_generate = True
@@ -135,73 +200,7 @@ numpydoc_validation_exclude = {  # set of regex
 copybutton_prompt_text = r">>> |\.\.\. |\$ "
 copybutton_prompt_is_regexp = True
 
-
-# The suffix(es) of source filenames.
-# You can specify multiple suffix as a list of string:
-#
-# source_suffix = ['.rst', '.md']
-source_suffix = ".rst"
-
-# The version info for the project you're documenting, acts as replacement for
-# |version| and |release|, also used in various other places throughout the
-# built documents.
-#
-# The short X.Y version.
-version = mne_icalabel.__version__
-# The full version, including alpha/beta/rc tags.
-release = version
-
[…]
