### 2022-08-20 e000fd8
remove badly defined source_suffix (rst is default anyway)
