+++ doc/conf.py
@@ -70,7 +70,6 @@ extensions = [
     "sphinx_issues",
 ]
 
-source_suffix = ".rst"
 templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "**.ipynb_checkpoints"]
 
