+++ doc/conf.py
@@ -82,6 +82,17 @@ nitpick_ignore = []
 # make `filter` a cross-reference to the Python function “filter”.
 default_role = "py:obj"
 
+# Resolve binder filepath_prefix. From the docs:
+# "A prefix to append to the filepath in the Binder links. You should use this
+# if you will store your built documentation in a sub-folder of a repository,
+# instead of in the root."
+# we will store dev docs in a `dev` subdirectory and all other docs in a
+# directory "v" + version_str. E.g., "v0.3"
+if "dev" in version:
+    filepath_prefix = "dev"
+else:
+    filepath_prefix = "v{}".format(version)
+
 # -- options for HTML output -------------------------------------------------
 
 # HTML options (e.g., theme)
