+++ doc/conf.py
@@ -129,18 +129,14 @@ html_sidebars = {
 }
 
 html_context = {
+    "pygment_light_style": "tango",
+    "pygment_dark_style": "native",
     "versions_dropdown": {
         "dev": "v0.4 (devel)",
         "stable": "v0.3",
         "v0.2": "v0.2",
         "v0.1": "v0.1",
     },
-}
-
-# variables to pass to HTML templating engine
-html_context = {
-    'pygment_light_style': 'tango',
-    'pygment_dark_style': 'native',
 }
 
 # -- autosummary -------------------------------------------------------------
