### 2022-08-22 mne-tools/mne-icalabel#105 @mscheltienne
Thank you for the review. PR #108 addresses this. Let us know if you think further changes are needed.
