### 2022-08-22 mne-tools/mne-icalabel#105 @adswa
Looks good to me, thx!
