### 2022-08-22 #4484 @adswa
mne-tools/mne-icalabel#105 is addressed, everything looks good to me, too
