### 2022-08-22 mne-tools/mne-icalabel#107 @adam2392
Thanks @adswa for all these comments. I agree w/ all of them and have added them into the paper.
