### 2022-08-22 #4484 @emdupre
Thank you for confirming your reviews, @adswa and @TomDonoghue !

@adam2392, I have a few last small comments and requests. Could you please :

- [ ] Correct the DOI for the Pion-Tonachini reference by removing the `https://doi.org/`prefix
- [ ] Add the DOI on the PyTorch citation. It is `10.48550/ARXIV.1912.01703`
- [ ] List the specific version of MNE-Python that you are compatible with ? From the toml file, it looks like this is >= 1.1, so you can directly update the first mention of the software; i.e., `(from v1.1; @Agramfort2013)`
- [ ] Add a 'the' on L18, so the phrase reads :
  > a software package compatible with the MNE-Python software toolkit
- [ ] On L40, update to current tense (as ICLabel is [still available](https://labeling.ucsd.edu/tutorial)):
  > ICLabel is a proposed statistical model
