### 2022-08-22 #4484 @emdupre
After making those edits, could you then please :

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare or an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata (e.g., the JOSS paper and software archive have the same title and author listing)
- [ ] List the DOI of the archived version here.

I can then move forward with accepting the submission 🚀
