### 2022-08-24 #4484 @emdupre
Thank you, @adam2392 !

My only remaining request is to update the Zenodo archive to correct the description's formatting; this can be done without minting a new DOI by [editing the metadata](https://help.zenodo.org/#technical) of the existing record. Please let me know if you have any issues with this !

Once this is done, I can recommend the submission for publication 🚀
