### 2022-08-24 #4484 @emdupre
That seems better, thank you @adam2392 !

I'm now happy to recommend MNE-ICALabel to the JOSS EIC team for publication—congratulations on this impressive effort !
