### 2022-08-24 #4484 @editorialbot
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
